% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrix.R
\name{activeScorePCAlist}
\alias{activeScorePCAlist}
\title{Compute PCA activation from a gene list.}
\usage{
activeScorePCAlist(exprMatrix, geneList, scale = FALSE, center = TRUE)
}
\arguments{
\item{exprMatrix}{A matrix of numeric with rows as features (in the RNA-Seq context, log counts).}

\item{geneList}{A named list containing character vectors of genes.}

\item{scale}{A character vector. The gene set where the activation score has to be computed. Must be a subset of `exprMatrix` row names.}

\item{center}{Logical. Subtract features by their average.}
}
\value{
A list of 2 element: `activScoreMat`: The activation score matrix. `contributionList`: A list of named numeric containing contributions of genes.
}
\description{
Compute PCA activation from a gene list.
}
\examples{
data("bulkLogCounts")
keggData<-getDBterms(rownames(bulkLogCounts),database = "kegg")
res<-activeScorePCAlist(bulkLogCounts,geneList = keggData$kegg[1:3])
activeScoreMat<-res$activScoreMat
contributionList<-res$contributionList
}
