% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrix.R
\name{activScorePCA}
\alias{activScorePCA}
\title{Compute the activation score of a gene set from 1st component of its PCA}
\usage{
activScorePCA(
  exprMatrix,
  genes,
  scale = FALSE,
  center = TRUE,
  returnContribution = FALSE
)
}
\arguments{
\item{exprMatrix}{A matrix of numeric with rows as features (in the RNA-Seq context, log counts).}

\item{genes}{A character vector. The gene set where the activation score has to be computed. Must be a subset of `exprMatrix` row names.}

\item{scale}{Logical. Divide features by their standard deviation.}

\item{center}{Logical. Subtract features by their average.}

\item{returnContribution}{Logical. Return list with activation score and contribution of genes to the activation score.}
}
\value{
A vector of numeric corresponding to activation scores, named by genes. If `returnContribution` return a list with activation scores and contributions of genes.
}
\description{
Compute the activation score of a gene set from 1st component of its PCA
}
\examples{
data("bulkLogCounts")
keggData<-getDBterms(rownames(bulkLogCounts),database = "kegg")
geneSet<-keggData$kegg$`hsa00190 Oxidative phosphorylation`
geneSet<-intersect(geneSet,rownames(bulkLogCounts))
activScorePCA(bulkLogCounts,genes = geneSet)
activScorePCA(bulkLogCounts,genes = geneSet,returnContribution = TRUE)
}
