% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrix.R
\name{adjMat2Pygraph}
\alias{adjMat2Pygraph}
\title{Python igraph object from an adjacency matrix.}
\usage{
adjMat2Pygraph(adjMat, mode = "directed", ...)
}
\arguments{
\item{adjMat}{An adjacency matrix of a graph.}

\item{mode}{A character, the mode to be used. Possible values are:
-"directed" - the graph will be directed and a matrix element gives the number of edges between two vertex.
-"undirected" - alias to "max" for convenience.
-"max" - undirected graph will be created and the number of edges between vertex i and j is max(A(i,j),A(j,i))
-"min" - like "max", but with min(A(i,j),A(j,i))
-"plus" - like "max", but with A(i,j)+A(j,i)
-"upper" - undirected graph with the upper right triangle of the matrix (including the diagonal)
-"lower" - undirected graph with the lower left triangle of the matrix (including the diagonal)}

\item{...}{Other parameters passed to the igraph python function `Weighted_Adjacency`.}
}
\value{
A python igraph object.
}
\description{
Python igraph object from an adjacency matrix.
}
\examples{
adjMat<-matrix(round(runif(25,min = 0,max = 1)),ncol = 5)
res<-adjMat2Pygraph(adjMat)
}
