% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoGparFontSizeMatrix}
\alias{autoGparFontSizeMatrix}
\title{Compute font size of col/rownames in complexHeatmap}
\usage{
autoGparFontSizeMatrix(n, ...)
}
\arguments{
\item{n}{Single numeric value. Number of column/rows}

\item{...}{Other parameters passed to gpar}
}
\value{
a gpar object
}
\description{
Compute font size of col/rownames in complexHeatmap
}
\examples{
autoGparFontSizeMatrix(5)
}
