% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{best.cutree}
\alias{best.cutree}
\title{Determine the best partition in a hierarchichal clustering}
\usage{
best.cutree(hc, min = 2, max = 20, loss = FALSE, graph = FALSE)
}
\arguments{
\item{hc}{A hclust object.}

\item{min}{Minimum number of class in the partition.}

\item{max}{Maximum number of class in the partition}

\item{loss}{Logical. Return the list of computed partition with their derivative loss.}

\item{graph}{Logical. Plot a graph of computed partition with their derivative loss.}
}
\value{
A single integer (best partition) or print a graph if \code{graph} or a vector of numeric if \code{loss}.
}
\description{
Determine the best partition in a hierarchichal clustering
}
\examples{
data(iris)
resClust<-hierarchicalClustering(iris[,c(1, 2, 3)],transpose = FALSE)
best.cutree(resClust,graph=TRUE)
best.cutree(resClust)
best.cutree(resClust, loss=TRUE)
cutree(resClust,k = best.cutree(resClust))
}
