% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oob-package.R
\docType{data}
\name{bulkLogCounts}
\alias{bulkLogCounts}
\title{bulkLogCounts}
\format{
A matrix containing 77 observations for 16959 features/genes
}
\description{
Log-expression count table for 77 samples sequenced from a bulk 3'digital RNA-Seq
protocol used in Kilens, Meistermann et al. 2018.
}
\references{
\url{https://doi.org/10.1038/s41467-017-02107-w}
}
\author{
Dimitri Meistermann
}
\keyword{data}
