% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDataframe.R
\name{chead}
\alias{chead}
\title{Print first columns and rows of a matrix/d}
\usage{
chead(x, n = 5)
}
\arguments{
\item{x}{A matrix or a dataframe}

\item{n}{number of columns/rows to print}
}
\value{
Print the result in the console.
}
\description{
Print first columns and rows of a matrix/d
}
\examples{
data("sampleAnnot")
chead(sampleAnnot)
}
