% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cn<-}
\alias{cn<-}
\title{colnames alias (setter)}
\usage{
cn(x) <- value
}
\arguments{
\item{x}{A matrix-like object.}

\item{value}{Either NULL or a character vector equal of length equal to the
appropriate dimension.}
}
\value{
The modified object.
}
\description{
colnames alias (setter)
}
\examples{
data(iris)
cn(iris)<-c("S.l","S.w", "P.l", "P.w", "Sp")
cn(iris)
}
