% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{computeActivationScore}
\alias{computeActivationScore}
\title{Compute the activation score of gene sets from an expression matrix.}
\usage{
computeActivationScore(
  expressionMatrix,
  corrIdGenes = NULL,
  scaleScores = FALSE,
  centerScores = TRUE,
  database = c("kegg", "reactom", "goBP", "goCC", "goMF"),
  maxSize = 500,
  minSize = 2,
  customAnnot = NULL,
  keggDisease = FALSE,
  species = "Human",
  db_terms = NULL,
  speciesData = NULL
)
}
\arguments{
\item{expressionMatrix}{An expression matrix (normalized log2(x+1) counts). Genes as rows and sample as columns. If `db_terms` is not given, must be named by gene symbols.}

\item{corrIdGenes}{Dataframe of gene ID correspondence where each column is a gene ID type. If not NULL `species` and `speciesData` arguments wont be used.}

\item{scaleScores}{Logical. Divide expression of gene by its standard deviation before doing the PCA.}

\item{centerScores}{Logical. Subtract mean to gene expression before doing the PCA.}

\item{database}{Which annotation database ? valid: database: kegg reactom goBP goCC goMF custom.}

\item{maxSize}{Maximum number of gene in each term.}

\item{minSize}{Minimum number of gene in each term.}

\item{customAnnot}{Custom annotation database, as a list of terms, each element contain a vector of gene symbols.}

\item{keggDisease}{Logical. Retain kegg disease term in kegg database?}

\item{species}{Character. Shortname of the species as described in `data("bods")`.}

\item{db_terms}{A list or NULL. A named list were each element is a database. Inside each database, a list terms, named by the term and containing gene vectors as gene symbols.}

\item{speciesData}{object returned by `getSpeciesData`. If not NULL `species` argument wont be used.}
}
\value{
A list where each element is a database of gene set given as input.
For each database, contain a list of activation score, with gene sets as rows and samples as columns ;
and the list of contribution (or weight) to activation score of each gene per gene set.
}
\description{
Perform a PCA for each gene set, from the matrix of [ genes from gene set × all samples ]. Return the first PCs as activations scores of the gene sets.
}
\examples{
data("bulkLogCounts")
keggDB<-getDBterms(rownames(bulkLogCounts),database = "kegg")
geneSetActivScore<-computeActivationScore(bulkLogCounts,db_terms = keggDB)
#same as
geneSetActivScore<-computeActivationScore(bulkLogCounts,database = "kegg")
}
