% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{convertColorAdd2Sub}
\alias{convertColorAdd2Sub}
\title{Convert color from additive to subtracting mixing}
\usage{
convertColorAdd2Sub(color, returnHex = TRUE)
}
\arguments{
\item{color}{A vector of 3 numeric containing rgb values (from 0 to 1), or a single character of the color name or hex code}

\item{returnHex}{Logical value indicating if the color should be returned as a hex code or as a vector of rgb values.}
}
\value{
a vector containing rgb values if returnHex=FALSE, otherwise a color as a character format.
}
\description{
Convert color from additive to subtracting mixing
}
\examples{
convertColorAdd2Sub(c(0,0,0))
convertColorAdd2Sub(c(0,0,0),returnHex = FALSE)

color<-rgb(red=.1,green=.1,blue=.1)
invertedColor<-convertColorAdd2Sub(color)

plotPalette(c(color,invertedColor))
}
