% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copyReadyVector}
\alias{copyReadyVector}
\title{Copy paste ready vector}
\usage{
copyReadyVector(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A string ready to be copied and embedded as R code.
}
\description{
Copy paste ready vector
}
\examples{
copyReadyVector(1:5)
}
