% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDataframe.R
\name{corrDist}
\alias{corrDist}
\title{Compute correlation distance}
\usage{
corrDist(x, method = "pearson")
}
\arguments{
\item{x}{A matrix of numeric. The function will compute the distance between the rows (same as \code{dist}).}

\item{method}{Character. Name of the method to compute correlation. See \code{method} argument from \code{cor}.}
}
\value{
An object of class "dist".
}
\description{
Compute correlation distance
}
\examples{
data("iris")
corrDist(t(iris[,seq_len(3)]))
}
