% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDataframe.R
\name{corrDistBicor}
\alias{corrDistBicor}
\title{Compute bicorrelation distance}
\usage{
corrDistBicor(x)
}
\arguments{
\item{x}{A matrix of numeric. The function will compute the distance between the rows (same as `dist`).}
}
\value{
An object of class "dist".
}
\description{
The bicorrelation distance is a metric used by WGCNA. In theory it is better to describe coexpression than Pearson's correlation.
}
\examples{
data("iris")
corrDistBicor(t(iris[,1:3]))
}
