% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDataframe.R
\name{cosineDist}
\alias{cosineDist}
\title{Compute cosine distance}
\usage{
cosineDist(x)
}
\arguments{
\item{x}{A matrix of numeric. The function will compute the distance between the rows (same as \code{dist}).}
}
\value{
An object of class "dist".
}
\description{
Compute cosine distance
}
\examples{
data("iris")
cosineDist(t(iris[,1:3]))
}
