% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDataframe.R
\name{covDist}
\alias{covDist}
\title{Compute covariance distance}
\usage{
covDist(x)
}
\arguments{
\item{x}{A matrix of numeric. The function will compute the distance between the rows (same as `dist`).}
}
\value{
An object of class "dist".
}
\description{
Compute covariance distance
}
\examples{
data("iris")
covDist(t(iris[,1:3]))
}
