% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrix.R
\name{CPM}
\alias{CPM}
\title{Count Per Million normalization}
\usage{
CPM(data, sce_assay = 1)
}
\arguments{
\item{data}{A raw count table with rows as genes
Can also be a \code{SummarizedExperiment} or \code{SingleCellExperiment} object.}

\item{sce_assay}{Integer or character,
if \code{data} is a \code{SummarizedExperiment} related object,
the assay name to use.}
}
\value{
A normalized count table.
If \code{data} is a \code{SummarizedExperiment} related object,
the function will add the normalized
count to the \code{SummarizedExperiment} object \code{cpm}.
}
\description{
Count Per Million normalization
}
\examples{
data("geneLengthGRCh38")
countMat <-
    sapply(vector("numeric", length = length(geneLengthGRCh38)), function(x){
        MASS::rnegbin(10, theta = abs(rnorm(1, mean = 10, sd = 20)),
                     mu = abs(rnorm(1, mean = 10, sd = 20)))
    }) |> t()
    rownames(countMat)<-names(geneLengthGRCh38)
chead(CPM(countMat))
sce <- SingleCellExperiment(assays = list(counts = countMat))
chead(assay(CPM(sce),"cpm"))
}
