% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cv2}
\alias{cv2}
\title{Coefficient of variation of squared mean and sd}
\usage{
cv2(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
A single numeric value
}
\description{
Coefficient of variation of squared mean and sd
}
\examples{
cv2(c(1,2,3,4))
}
