% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oob-package.R
\docType{data}
\name{DEgenesPrime_Naive}
\alias{DEgenesPrime_Naive}
\title{DEgenesPrime_Naive}
\format{
A data.frame containing 16959 observations for 7 features
}
\description{
Results of the differential expression analysis between primed and naive
human pluripotent stem cells in Kilens, Meistermann et al. 2018.
}
\references{
\url{https://doi.org/10.1038/s41467-017-02107-w}
}
\author{
Dimitri Meistermann
}
\keyword{data}
