% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDataframe.R
\name{drawSamplePerGroup}
\alias{drawSamplePerGroup}
\title{Draw n samples from each population and return it has a named vector}
\usage{
drawSamplePerGroup(
  sampleNames,
  group,
  maxDrawSize = NULL,
  minDrawSize = NULL,
  replace = FALSE
)
}
\arguments{
\item{sampleNames}{A character vector of the name of the samples.}

\item{group}{A factor or a character vector of the same length as \code{sampleNames}. Describe the population of each sample.}

\item{maxDrawSize}{Maximum number of observation to draw per group.}

\item{minDrawSize}{Minimum number of observation to draw per group.}

\item{replace}{Logical. Should the sample be drawn with replacement.}
}
\value{
A named vector with \code{sampleNames} with the population of each sample.
}
\description{
Draw n samples from each population and return it has a named vector
}
\examples{
sampleNames<-paste0("sample",1:20)
group<-c(rep("A",5),rep("B",12),rep("C",3))
drawSamplePerGroup(sampleNames,group)
drawSamplePerGroup(sampleNames,group,minDrawSize=6)
drawSamplePerGroup(sampleNames,group,maxDrawSize=2)

# If minDrawSize > number of sample in a group and replace=TRUE, samples
# will be drawn with replacement in this group
drawSamplePerGroup(sampleNames,group,minDrawSize=6, replace=TRUE)
}
