% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichSetIntersection}
\alias{enrichSetIntersection}
\title{Compute p-value and other metric for enrichment of set intersection.}
\usage{
enrichSetIntersection(intersectionSize, setSizes, universeSize)
}
\arguments{
\item{intersectionSize}{Numeric. Number of common element between sets.}

\item{setSizes}{Vector of numeric. Number of element in each set.}

\item{universeSize}{Total number of element in the universe.}
}
\value{
A list of values:
- observed: observed number of elements in intersection (same value as `intersectionSize`)
- expected: expected number of elements in intersection.
- log2OE: log2 of observed/expected ratio
- pval: pvalue (upper tail) where NULL hypotheses is that observed intersection is from the distribution of intersection size with p = nExpected / nUniverse.
}
\description{
Compute p-value and other metric for enrichment of set intersection.
}
\details{
If two sets, perform an hypergeometric test, if more the p-value is computed from binomial distribution.
}
\examples{
setSizes<-c(150,200,250)
universeSize<-10000
intersectionSize<-5
enrichSetIntersection(intersectionSize, setSizes, universeSize)
}
