% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{exportEnrich}
\alias{exportEnrich}
\title{Export enrichment results with the gene list associated to each term/row.}
\usage{
exportEnrich(
  enrichResults,
  file,
  quote = FALSE,
  sep = "\\t",
  col.names = TRUE,
  row.names = FALSE,
  geneCol = "genes",
  ...
)
}
\arguments{
\item{enrichResults}{Dataframe, usually from Enrich functions (for example `enrich.ora`) with `returnGenes=TRUE`.}

\item{file}{Path to file to write.}

\item{quote}{a logical value (TRUE or FALSE) or a numeric vector. If TRUE, any character or factor columns will be surrounded by double quotes. If a numeric vector, its elements are taken as the indices of columns to quote. In both cases, row and column names are quoted if they are written. If FALSE, nothing is quoted.}

\item{sep}{Field separator.}

\item{col.names}{Either a logical value indicating whether the column names of x are to be written along with x, or a character vector of column names to be written. See the section on ‘CSV files’ for the meaning of col.names = NA.}

\item{row.names}{Either a logical value indicating whether the row names of x are to be written along with x, or a character vector of row names to be written.}

\item{geneCol}{The column that contain the list of gene of the term.}

\item{...}{Other parameters passed to write.table.}
}
\description{
Export enrichment results with the gene list associated to each term/row.
}
