% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{extendColorPalette}
\alias{extendColorPalette}
\title{Add or remove colors to an existing palette by interpolation}
\usage{
extendColorPalette(
  n,
  colors = c("#E52421", "#66B32E", "#2A4B9B", "#6EC6D9", "#F3E600", "#A6529A", "#7C1623",
    "#006633", "#29235C", "#0084BC", "#E6007E", "#F49600", "#E3E3E3", "#626F72",
    "#040505", "#E74B65", "#95B37F", "#683C11", "#F8BAA0", "#DD8144"),
  sortColorIn = FALSE,
  sortColorOut = FALSE
)
}
\arguments{
\item{n}{Number of returned colors.}

\item{colors}{Vector of colors.}

\item{sortColorIn}{Order color vector by similarity before the interpolation.}

\item{sortColorOut}{Order color vector by dissimilarity after the interpolation.}
}
\value{
Vector of colors.
}
\description{
Add or remove colors to an existing palette by interpolation
}
\examples{
extendColorPalette(9,  colors=c("red","green","blue")) |> plotPalette()
extendColorPalette(9,  colors=c("red","green","blue"), sortColorIn=TRUE, sortColorOut=TRUE) |> plotPalette()
}
