% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDataframe.R
\name{factorAsStrings}
\alias{factorAsStrings}
\title{Convert factors in a dataframe to strings}
\usage{
factorAsStrings(dt)
}
\arguments{
\item{dt}{A dataframe containing factor column}
}
\value{
A dataframe where all the factor column have been converted to srings
}
\description{
Convert factors in a dataframe to strings
}
\examples{
a<-data.frame(x=factor(c("a","a","b")),y=1:3)
b<-factorAsStrings(a)
b$x
}
