% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{factorToVectorList}
\alias{factorToVectorList}
\title{Convert a named factor vector to a list}
\usage{
factorToVectorList(factorValues, factorNames = NULL)
}
\arguments{
\item{factorValues}{A vector of factor. It has to be named if `factorNames=NULL`.}

\item{factorNames}{A character vector for providing the names separately.}
}
\value{
A list. Each element is named by a factor level of `factorValues`, and contains the provided names that had this level has a value.
}
\description{
Convert a named factor vector to a list
}
\examples{
x<-factor(c("a","a","b","b","c","c","c"))
names(x)<-paste0("x",1:7)
factorToVectorList(x)

}
\seealso{
VectorListToFactor
}
