% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{fastRead}
\alias{fastRead}
\title{Load quickly a text file to a dataframe/matrix}
\usage{
fastRead(
  fileName,
  sep = "\\t",
  row.names = 1,
  as.matrix = FALSE,
  stringsAsFactors = FALSE,
  ...
)
}
\arguments{
\item{fileName}{Path to the file.}

\item{sep}{Separator ("\\t"=tab-separated values).}

\item{row.names}{Column that contains the row names,set to NULL for not
attributing row names.}

\item{as.matrix}{Return a dataframe if FALSE, a matrix if TRUE.}

\item{stringsAsFactors}{Convert automatically strings to factors if TRUE.}

\item{...}{Other parameters passed to data.table::fread.}
}
\value{
A dataframe or a matrix.
}
\description{
Load quickly a text file to a dataframe/matrix
}
\examples{
data("bulkLogCounts")
tf <- tempfile(fileext = ".tsv")
fastWrite(bulkLogCounts,tf)
fastRead(tf)
}
\seealso{
fastWrite
}
