% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{fastWrite}
\alias{fastWrite}
\title{Write quickly a text file from dataframe/matrix}
\usage{
fastWrite(
  x,
  fileName = "default.tsv",
  headRow = "Name",
  row.names = TRUE,
  col.names = TRUE,
  dec = ".",
  sep = "\\t",
  ...
)
}
\arguments{
\item{x}{matrix/dataframe to save.}

\item{fileName}{Path to the file to write.}

\item{headRow}{Name in the header of the rownames column.}

\item{row.names}{Whether or not to save rownames.}

\item{col.names}{Whether or not to save colnames.}

\item{dec}{Character that serve as a decimal separator.}

\item{sep}{Separator ("\\t"=tab-separated values).}

\item{...}{Other parameters passed to data.table::fwrite.}
}
\value{
NULL, write a file.
}
\description{
Write quickly a text file from dataframe/matrix
}
\examples{
data("bulkLogCounts")
tf <- tempfile(fileext = ".tsv")
fastWrite(bulkLogCounts,tf)
}
\seealso{
fastRead
}
