% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{fcsScoreDEgenes}
\alias{fcsScoreDEgenes}
\title{Compute a "interest score" for a set of genes. Useful for GSEA.}
\usage{
fcsScoreDEgenes(genes, pvalues, logFoldChanges, logPval = FALSE)
}
\arguments{
\item{genes}{A vector of gene names}

\item{pvalues}{A vector of numeric corresponding to p-values.}

\item{logFoldChanges}{A vector of numeric corresponding to Log2(Fold-change)}

\item{logPval}{Logical. Compute the p-value score as `-log10(pval)` instead of `1-pval`.}
}
\value{
A vector of numeric corresponding to interest scores, named by genes.
}
\description{
Compute a "interest score" for a set of genes. Useful for GSEA.
}
\examples{
data("DEgenesPrime_Naive")
fcsScore<-fcsScoreDEgenes(rownames(DEgenesPrime_Naive),DEgenesPrime_Naive$pvalue,DEgenesPrime_Naive$log2FoldChange)
}
