% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{filledDoubleArrow}
\alias{filledDoubleArrow}
\title{Plot a double arrow in a grid plot.}
\usage{
filledDoubleArrow(
  x = 0,
  y = 0,
  width = 1,
  height = 1,
  just = c("left", "bottom"),
  gp = gpar(col = "black"),
  ...
)
}
\arguments{
\item{x}{Numeric. X coordinate of the arrow.}

\item{y}{Numeric. Y coordinate of the arrow.}

\item{width}{Numeric. X coordinate of the arrow.}

\item{height}{Numeric. X coordinate of the arrow.}

\item{just}{A string or numeric vector specifying the justification of the
viewport relative to its (x, y) location. If there are two values, the
first value specifies horizontal justification and the second value
specifies vertical justification. Possible string values are: "left",
"right", "centre", "center", "bottom", and "top". For numeric values, 0
means left alignment and 1 means right alignment.}

\item{gp}{An object of class "gpar", typically the output from a call to the
function gpar. This is basically a list of graphical parameter settings}

\item{...}{Other parameters passed to pushViewport.}
}
\value{
Plot in the current graphical device.
}
\description{
Plot a double arrow in a grid plot.
}
\examples{
filledDoubleArrow()
oobqplot(seq_len(5))
filledDoubleArrow()
}
