% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oob-package.R
\docType{data}
\name{geneLengthGRCh38}
\alias{geneLengthGRCh38}
\title{geneLengthGRCh38}
\format{
A vector of 25017 numeric named with genes.
}
\description{
Gene length table for 25017 genes from the GRCh38 annotation
used in Kilens, Meistermann et al. 2018.
}
\references{
\url{https://doi.org/10.1038/s41467-017-02107-w}
}
\author{
Dimitri Meistermann
}
\keyword{data}
