% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{getDBterms}
\alias{getDBterms}
\title{Download database of term/pathway for enrichment analyses.}
\usage{
getDBterms(
  geneSym,
  geneEntrez = NULL,
  corrIdGenes = NULL,
  speciesData = NULL,
  database = c("kegg", "reactom", "goBP", "goCC", "goMF"),
  customAnnot = NULL,
  keggDisease = FALSE,
  species = "Human",
  returnGenesSymbol = TRUE
)
}
\arguments{
\item{geneSym}{A vector of gene symbols.}

\item{geneEntrez}{NULL or a vector of gene Entrez ID.}

\item{corrIdGenes}{Dataframe of gene ID correspondence where each column is a gene ID type. If not NULL `species` and `speciesData` arguments wont be used.}

\item{speciesData}{object returned by `getSpeciesData`. If not NULL `species` argument wont be used.}

\item{database}{Which annotation database ? valid: database: kegg reactom goBP goCC goMF custom.}

\item{customAnnot}{Custom annotation database, as a list of terms, each element contain a vector of gene symbols.}

\item{keggDisease}{Logical. Retain kegg disease term in kegg database?}

\item{species}{Character. Shortname of the species as described in `data("bods")`.}

\item{returnGenesSymbol}{Logical, return gene symbol in each term instead of Entrez ID.}
}
\value{
A list where each element is a database that contain a list of term with the associated gene symbols.
}
\description{
Download database of term/pathway for enrichment analyses.
}
\examples{
data("bulkLogCounts")
enrichDBs<-getDBterms(rownames(bulkLogCounts),species="Human",database=c("kegg","reactom"))
}
