% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers_diffExpr.R
\name{getMarkerGLMnet}
\alias{getMarkerGLMnet}
\title{Compute a matrix of coef describing best gene marker per group of samples
from a GLM net regression}
\usage{
getMarkerGLMnet(data, group, transpose = TRUE, sce_assay = 1)
}
\arguments{
\item{data}{A matrix of numeric with rows as features (in the
RNA-Seq context, log data).
Can also be a \code{SummarizedExperiment} or \code{SingleCellExperiment} object.}

\item{group}{A feature of factor/character, same length as number of sample.
Describe group of each sample.}

\item{transpose}{If TRUE, the input data is transposed before processing.
Default is TRUE (feature as rows, samples as columns).}

\item{sce_assay}{Integer or character, if \code{data} is a
\code{SummarizedExperiment} related object, the assay name to use.}
}
\value{
A matrix containing each gene coefficient for each group.
If \code{data} is a \code{SummarizedExperiment} related object
the function will add the marker metrics to \code{rowData}.
}
\description{
Compute a matrix of coef describing best gene marker per group of samples
from a GLM net regression
}
\examples{
data("bulkLogCounts")
data("sampleAnnot")
res <- getMarkerGLMnet(bulkLogCounts,sampleAnnot$culture_media)
sce <- SingleCellExperiment(assays = list(counts = bulkLogCounts))
res <- getMarkerGLMnet(sce,sampleAnnot$culture_media)
}
