% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRandState}
\alias{getRandState}
\title{Get the precise random seed state}
\usage{
getRandState()
}
\value{
A vector of integers representing the random seed state
}
\description{
Get the precise random seed state
}
\examples{
randSate<-getRandState()
rnorm(5)
setRandState(randSate)
rnorm(5)
}
\seealso{
setRandState
}
