% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{getSpeciesData}
\alias{getSpeciesData}
\title{Return gene id correspondance, GO species code and KEGG species code}
\usage{
getSpeciesData(sample.species = "Human", updateSpeciesPackage = FALSE)
}
\arguments{
\item{sample.species}{Character. Shortname of the species as described in `data("bods")`.}

\item{updateSpeciesPackage}{Logical. Download or update automatically the annotation org.db package corresponding to the species.}
}
\value{
A list describing specific data for the species (gene IDs, annotation package...).
}
\description{
Return gene id correspondance, GO species code and KEGG species code
}
\examples{
library(gage)
data(bods)
bods
getSpeciesData("Human")
getSpeciesData("Mouse")
}
