% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggBorderedFactors}
\alias{ggBorderedFactors}
\title{Add line break between factors and remove line in the middle if x axis is discrete.}
\usage{
ggBorderedFactors(gg, borderSize = 0.75, borderColor = "grey75")
}
\arguments{
\item{gg}{ggplot object}

\item{borderSize}{Single numeric value. Size width of the line.}

\item{borderColor}{Single character value. Color of the line.}
}
\value{
A ggplot objet.
}
\description{
Add line break between factors and remove line in the middle if x axis is discrete.
}
\examples{
g<-ggplot(data.frame(x=c("A","A","B","B","B","C")),aes(x=x))+geom_bar()
g
ggBorderedFactors(g)
ggBorderedFactors(g,borderColor="white",borderSize=1.5)
}
