% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggplotBreak}
\alias{ggplotBreak}
\title{Create breaks for a custom ggplot scale}
\usage{
ggplotBreak(x, scale, m = 5)
}
\arguments{
\item{x}{A vector of numeric.}

\item{scale}{A scale object as produced by scales::trans_new}

\item{m}{The number of desired breaks}
}
\value{
A vector of numeric for the major breaks
}
\description{
Create breaks for a custom ggplot scale
}
\examples{
data("DEgenesPrime_Naive")
custom_scale<-scales::trans_new(name = "invLog10",
    transform = function(x) -log10(x),
    inverse = function(x) 10^(-x), domain = c(0, Inf))
breaks <- ggplotBreak(DEgenesPrime_Naive$pvalue, custom_scale)
ggplot(DEgenesPrime_Naive, aes(x=log2FoldChange, y=pvalue)) +
    geom_point() +
    scale_y_continuous(trans = custom_scale, breaks = breaks)
}
