% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggplotColours}
\alias{ggplotColours}
\title{Default Color scale of ggplot2}
\usage{
ggplotColours(n = 6, h = c(0, 360) + 15)
}
\arguments{
\item{n}{Single integer. Number of colors to be returned.}

\item{h}{numeric vector of 2 value. Hue range (see \code{hcl})}
}
\value{
A character vector of colors.
}
\description{
Default Color scale of ggplot2
}
\examples{
ggplotColours()
ggplotColours(5)
ggplotColours(5,)
}
