% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gmean}
\alias{gmean}
\title{Compute geometrical mean}
\usage{
gmean(x, keepZero = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{keepZero}{get rid of 0 before computing.}
}
\value{
A single numeric value.
}
\description{
Compute geometrical mean
}
\examples{
gmean(c(1,2,3))
gmean(c(0,2,3),keepZero = TRUE)
gmean(c(0,2,3),keepZero = FALSE)
}
