% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{GSDA.HeatmapAnnot}
\alias{GSDA.HeatmapAnnot}
\title{Displaying most neg and pos gene contribution on a GSDA heatmap.}
\usage{
GSDA.HeatmapAnnot(
  contributions,
  maxGeneContribAtOneSide = 3,
  width = unit(3, "cm"),
  fontsizeFactor = 400
)
}
\arguments{
\item{contributions}{A list named by pathway. Contains vector of gene contribution to activation scores, named by genes.}

\item{maxGeneContribAtOneSide}{Integer. Number of best pos/neg rank displayed on the annotation}

\item{width}{Numeric. Width of the annotation heatmap.}

\item{fontsizeFactor}{Numeric. Font-size of gene names.}
}
\value{
A HeatmapAnnotation object. Genes on the left side of the vertical line are contributing negatively to the activation score, and positively on the right side.
}
\description{
Displaying most neg and pos gene contribution on a GSDA heatmap.
}
\examples{
data("bulkLogCounts")
data("sampleAnnot")

keggDB<-getDBterms(rownames(bulkLogCounts),database = "kegg")
geneSetActivScore<-computeActivationScore(bulkLogCounts,db_terms = keggDB)
resGSDA<-GSDA(geneSetActivScore = geneSetActivScore,colData = sampleAnnot,contrast = c("culture_media","T2iLGO","KSR+FGF2"),db_terms =  keggDB)
bestPathay<-whichTop(resGSDA$padj,top = 30,decreasing = FALSE)

heatmap.DM(geneSetActivScore$kegg$eigen[bestPathay,],midColorIs0 = TRUE,center=FALSE,
	name = "Activation score",preSet = NULL,colData = sampleAnnot["culture_media"],
	right_annotation=rowAnnotation("gene contribution" =
		GSDA.HeatmapAnnot(contributions = geneSetActivScore$kegg$contribution[bestPathay],width = unit(12,"cm"),fontsizeFactor = 300)
	),
	row_names_side ="left",row_dend_side ="left",
	row_names_max_width = unit(8, "inches"),autoFontSizeRow=FALSE,row_names_gp=gpar(fontsize=1/length(bestPathay)*300)
)
}
