% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{GSDA}
\alias{GSDA}
\title{Gene Set Differential Activation (GSDA)}
\usage{
GSDA(
  geneSetActivScore = NULL,
  expressionMatrix = NULL,
  colData,
  contrast,
  corrIdGenes = NULL,
  database = c("kegg", "reactom", "goBP", "goCC", "goMF"),
  maxSize = 500,
  minSize = 2,
  customAnnot = NULL,
  keggDisease = FALSE,
  species = "Human",
  db_terms = NULL,
  speciesData = NULL
)
}
\arguments{
\item{geneSetActivScore}{A list of database with an element "eigen" containing the matrix of gene set activation score (see what `computeActivationScore` returns).
If NULL, this is computed automatically from the `expressionMatrix` and the gene set database given via `db_terms` or requested via `database`.}

\item{expressionMatrix}{An expression matrix (normalized log2(x+1) counts). Genes as rows and sample as columns. If `db_terms` is not given, must be named by gene symbols.}

\item{colData}{An annotation dataframe. Each column is a feature, each row a sample. Same number of samples than in `expressionMatrix`.}

\item{contrast}{A vector of 3 character.
1. Name of the experimental variable that have to be used for differential activation. Must be a column name of `colData`.
2. Condition considered as the reference.
3. Condition considered as the target group.}

\item{corrIdGenes}{Dataframe of gene ID correspondence where each column is a gene ID type. If not NULL `species` and `speciesData` arguments wont be used.}

\item{database}{Which annotation database ? valid: database: kegg reactom goBP goCC goMF custom.}

\item{maxSize}{Maximum number of gene in each term.}

\item{minSize}{Minimum number of gene in each term.}

\item{customAnnot}{Custom annotation database, as a list of terms, each element contain a vector of gene symbols.}

\item{keggDisease}{Logical. Retain kegg disease term in kegg database?}

\item{species}{Character. Shortname of the species as described in `data("bods")`.}

\item{db_terms}{A list or NULL. A named list were each element is a database. Inside each database, a list terms, named by the term and containing gene vectors as gene symbols.}

\item{speciesData}{object returned by `getSpeciesData`. If not NULL `species` argument wont be used.}
}
\value{
A dataframe with the following columns:
- term: name of the term/gene set
- baseMean: mean of activation score in the gene set
- sd: standard deviation  in the gene set
- log2FoldChange: Log(Log Fold Change) of activation score between the two tested groups.
- pval: an enrichment p-value
- padj: a BH-adjusted p-value
- database: origin of the gene set
- size: number of gene in the term after removing genes not present.
}
\description{
Gene Set Differential Activation (GSDA)
}
\examples{
data("bulkLogCounts")
data("sampleAnnot")

keggDB<-getDBterms(rownames(bulkLogCounts),database = "kegg")
geneSetActivScore<-computeActivationScore(bulkLogCounts,db_terms = keggDB)
resGSDA<-GSDA(geneSetActivScore = geneSetActivScore,colData = sampleAnnot,contrast = c("culture_media","T2iLGO","KSR+FGF2"),db_terms =  keggDB)

#or
resGSDA<-GSDA(expressionMatrix = bulkLogCounts,colData = sampleAnnot,contrast = c("culture_media","T2iLGO","KSR+FGF2"),database = "kegg")
}
