% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{htMarker}
\alias{htMarker}
\title{Plot best marker per group on a tidy Heatmap}
\usage{
htMarker(
  countTable,
  group,
  markerData,
  colData = NULL,
  topn = 5,
  maxDrawSize = NULL,
  minDrawSize = NULL,
  replace = FALSE,
  returnHeatmap = FALSE,
  show_column_names = FALSE,
  ...
)
}
\arguments{
\item{countTable}{A matrix of numeric with samples as columns (in the RNA-Seq
context, log counts)}

\item{group}{A feature of factor/character, same length as number of sample.
Describe group of each sample (for example clusters).}

\item{markerData}{A matrix describing marker scores for each group.}

\item{colData}{A vector of factor, character, numeric or logical. Or, a
dataframe of any of these type of value. The annotation that will be
displayed on the heatmap.}

\item{topn}{Number of marker to plot, rankes from the best}

\item{maxDrawSize}{Maximum number of observation to draw per group.}

\item{minDrawSize}{Minimum number of observation to draw per group.}

\item{replace}{Logical. Should the sample be drawn with replacement.}

\item{returnHeatmap}{Logical, return the plot as a Heatmap object or print it
in the current graphical device.}

\item{show_column_names}{Whether show column names.}

\item{...}{Arguments passed to \code{heatmap.DM}}
}
\value{
A Heatmap object if \code{returnHeatmap} or print the Heatmap in the
current graphical device.
}
\description{
Plot best marker per group on a tidy Heatmap
}
\details{
Draw the same number of observation from each condition/group, and
take the top n marker per group.
The Heatmap is sliced by group of samples, and by their respective markers.
}
\examples{
data("bulkLogCounts")
data("sampleAnnot")
markerData <- getMarkers(bulkLogCounts,sampleAnnot$culture_media,
    BPPARAM=BiocParallel::SnowParam(1))
htMarker(bulkLogCounts,  group=sampleAnnot$culture_media,
    markerData=extractFeatureMarkerData(markerData),
    colData=sampleAnnot[c("line","passage")])
}
\seealso{
\code{\link[=heatmap.DM]{heatmap.DM()}}
}
