% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oob-package.R
\docType{data}
\name{humanGeneIDtable}
\alias{humanGeneIDtable}
\title{humanGeneIDtable}
\format{
A data frame containing 199476 rows for 3 columns
}
\description{
Gene ID table corresponance between ENTREZ ID, gene SYMBOL and ENSEMBL ID
from the GRCh38 annotation. Used in Kilens, Meistermann et al. 2018.
}
\references{
\url{https://doi.org/10.1038/s41467-017-02107-w}
}
\author{
Dimitri Meistermann
}
\keyword{data}
