% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inter}
\alias{inter}
\title{intersect alias}
\usage{
inter(x, y, ...)
}
\arguments{
\item{x}{A vector-like object.}

\item{y}{A vector-like object.}

\item{...}{Additional arguments passed to \code{intersect}.}
}
\description{
intersect alias
}
