% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDataframe.R
\name{intersectionEnrichment}
\alias{intersectionEnrichment}
\title{Enrichment values of intersection sets}
\usage{
intersectionEnrichment(isInGroupMat)
}
\arguments{
\item{isInGroupMat}{A matrix of integer}
}
\value{
A single numeric. This enrichment value is > 1 if the sets have more values in common than expected. Less than expected if enrichment value < 1.
}
\description{
Enrichment values of intersection sets
}
\examples{
g1<-letters[1:5]
g2<-letters[2:10]
isInGroupMat<-matrixFromDimnames(letters,c("g1","g2"))
isInGroupMat[g1,"g1"]<-1
isInGroupMat[g2,"g2"]<-1

#another to way compute isInGroupMat
isInGroupMat<-ComplexHeatmap::list_to_matrix(lt = list(g1=g1,g2=g2),universal_set = letters)
intersectionEnrichment(isInGroupMat)
}
