% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{linearScale}
\alias{linearScale}
\title{Transform a range of value to another by a linear relationship.}
\usage{
linearScale(vals, newRange, returnFunction = TRUE)
}
\arguments{
\item{vals}{A numeric vector. Values to be transposed in the new range.}

\item{newRange}{A vector of two numeric values: the new minimum and maximum.}

\item{returnFunction}{Logical. Return the linear scale as a function instead of the transposed values in a new scale. If set to \code{TRUE}, \code{vals} argument can be also a vector of 2 numeric corresponding to the minimum and maximum of the old range.}
}
\value{
A vector of value or a function if \code{returnFunction=TRUE}.
}
\description{
Similar to the javascript function \code{d3.scaleLinear()}.
}
\examples{
oldValues<-seq_len(10)
linearScale(oldValues,c(0,1),returnFunction = FALSE)
scaleFun<-linearScale(c(1,10),c(0,1),returnFunction = TRUE)
scaleFun(oldValues)
}
