% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDataframe.R
\name{matrixFromDimnames}
\alias{matrixFromDimnames}
\title{Create empty matrix from a column and rownames}
\usage{
matrixFromDimnames(row, col, value = 0)
}
\arguments{
\item{row}{A vector of character.}

\item{col}{A vector of character.}

\item{value}{Default value for each element of the matrix.}
}
\value{
A matrix. Size of the matrix will be `length(row)*length(col)`.
}
\description{
Create empty matrix from a column and rownames
}
\examples{
matrixFromDimnames(1:10,letters[1:5])
}
