% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Mode}
\alias{Mode}
\title{Compute the mode of a distribution.}
\source{
https://github.com/benmarwick/LaplacesDemon/blob/master/R/Mode.R
}
\usage{
Mode(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A single numeric value.
}
\description{
Compute the mode of a distribution.
}
\examples{
Mode(c(seq_len(10),3))

}
