% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrix.R
\name{multiLinearModel}
\alias{multiLinearModel}
\title{Test a linear model on each gene following an experimental design.}
\usage{
multiLinearModel(exprData, colData, designFormula, contrast)
}
\arguments{
\item{exprData}{A matrix of numeric with rows as features (in the RNA-Seq context, log counts).}

\item{colData}{A dataframe of feature (numeric or factor) with rows as samples. Must have the same number of samples than exprData}

\item{contrast}{A vector of 3 character.
1. Name of the experimental variable that have to be used for differential activation. Must be a column name of `colData`.
2. Condition considered as the reference.
3. Condition considered as the target group.}
}
\value{
A dataframe with the following columns:
- baseMean: mean
- log2FoldChange: Log(Log Fold Change)  between the two tested groups.
- pval: an enrichment p-value
- padj: a BH-adjusted p-value
}
\description{
Test a linear model on each gene following an experimental design.
}
\examples{
data("bulkLogCounts")
data("sampleAnnot")
res<-multiLinearModel(bulkLogCounts,colData = sampleAnnot,contrast = c("culture_media","T2iLGO","KSR+FGF2"))
}
