% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple ggplot one one page}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{Several plot from ggplot. Each given as an argument.}

\item{plotlist}{Plots given as a list. Each element contains a plot from grid or inherited package.}

\item{cols}{Number of columns. If layout is NULL, then use 'cols' to determine layout}

\item{layout}{A 2D matrix of numeric value (from one the the number of plot) indicating the layout to be plotted. Matrix()}
}
\description{
Multiple ggplot one one page
}
\examples{
p1<-ggplot(data = data.frame(x=1:5,y=1:5),aes(x=x,y=y))+geom_point()+ggtitle("p1")
p2<-qplot(c("a","a","b","b","b","c"))+ggtitle("p2")
p3<-qplot(c("a","a","b","b","b"))+ggtitle("p3")
p4<-qplot(c("a","b","c","d"))+ggtitle("p4")

multiplot(p1,p2,p3,p4)
plotList<-list(p1,p2,p3,p4)
multiplot(plotlist = plotList)
multiplot(plotlist = plotList,cols = 2)

layout<-matrix(data = c(
	4,3,
	2,1
),ncol = 2,byrow = TRUE)
multiplot(plotlist = plotList,layout = layout)
layout<-matrix(data = c(
	1,2,3,
	4,0,0
),ncol = 3,byrow = TRUE)
multiplot(plotlist = plotList,layout = layout)
}
