% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReductions.R
\name{NMDS}
\alias{NMDS}
\title{Non-metric Multidimensional Scaling (dimension reduction)}
\usage{
NMDS(
  data,
  transpose = TRUE,
  scale = FALSE,
  center = FALSE,
  metric = dist,
  ndim = 2,
  maxit = 100,
  sce_assay = "logcounts"
)
}
\arguments{
\item{data}{A matrix of numeric (in the RNA-Seq context, log counts).
Can also be a SingleCellExperiment object.}

\item{transpose}{Logical. If \code{transpose}, samples are columns and features
are rows.}

\item{scale}{Logical. Divide features by their standard deviation.}

\item{center}{Logical. Subtract features by their average.}

\item{metric}{A function that return a object of class "dist".}

\item{ndim}{Integer. Dimensions of the embedded space.}

\item{maxit}{The maximum number of iterations.}

\item{sce_assay}{Integer or character,
if \code{data} is a \code{SingleCellExperiment} object, the assay name to use.}
}
\value{
A matrix of coordinates with samples as rows, or a
\code{SingleCellExperiment} object if \code{data} is a \code{SingleCellExperiment.} with
the NMDS stored in a \code{reducedDim} slot.
}
\description{
Non-metric Multidimensional Scaling (dimension reduction)
}
\examples{
data("bulkLogCounts")
NMDSproj<-NMDS(bulkLogCounts)
proj2d(NMDSproj)
sce <- SingleCellExperiment(assays = list(logcounts = bulkLogCounts))
sce<-NMDS(sce)
proj2d(sce)
}
