% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrix.R
\name{normDeseq}
\alias{normDeseq}
\title{Quick DESeq2 normalization}
\usage{
normDeseq(countMatrix, sce_assay = 1)
}
\arguments{
\item{countMatrix}{A raw count table with rows as genes.
Can also be a \code{SummarizedExperiment} or \code{SingleCellExperiment}  object.}

\item{sce_assay}{Integer or character,
if \code{data} is a \code{SummarizedExperiment} related object,
the assay name to use.}
}
\value{
A normalized count table.
If \code{data} is a \code{SummarizedExperiment} related object, the function will
add the normalized count to
the \code{SummarizedExperiment} object as an assay named \code{deseq2norm}.
}
\description{
Quick DESeq2 normalization
}
\examples{
data("geneLengthGRCh38")
countMat <-
    sapply(vector("numeric", length = length(geneLengthGRCh38)), function(x){
        MASS::rnegbin(10, theta = abs(rnorm(1, mean = 10, sd = 20)),
        mu = abs(rnorm(1, mean = 10, sd = 20)))
    }) |> t()
rownames(countMat)<-names(geneLengthGRCh38)
normDeseq(countMat)
sce <- SingleCellExperiment(assays = list(counts = countMat))
assay(normDeseq(sce),"deseq2norm") |> chead()
}
