% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{oobColors}
\alias{oobColors}
\title{Colors for a qualitative scale}
\usage{
oobColors(n = 20)
}
\arguments{
\item{n}{Number of wanted colors}
}
\value{
A vector of colors
}
\description{
Colors for a qualitative scale
}
\examples{
oobColors() |> plotPalette()
oobColors(n=5) |> plotPalette()
oobColors(n=40) |> plotPalette()
}
