% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrix.R
\name{oobFastMNN}
\alias{oobFastMNN}
\title{Execute a fastMNN and rescale the counts.}
\usage{
oobFastMNN(
  logCounts,
  batch,
  k,
  returnRescale = TRUE,
  sce_assay = "logcounts",
  ...
)
}
\arguments{
\item{logCounts}{A matrix of numeric (in the RNA-Seq context, log counts).
Can also be a \code{SummarizedExperiment} or \code{SingleCellExperiment} object.}

\item{batch}{A vector or factor specifying the batch of origin for all cells}

\item{k}{An integer scalar specifying the number of nearest neighbors to
consider when identifying MNNs.}

\item{returnRescale}{Logical. Use reScale on the output so the dynamic range
after batch correction of genes is the same than before.}

\item{sce_assay}{Integer or character, if \code{data}
is a \code{SummarizedExperiment} related object, the assay name to use.}

\item{...}{Other parameters passed to fastMNN.}
}
\value{
A matrix of corrected count table.
}
\description{
Use fastMNN fron the package batchelor, but return a matrix.
Rescale the counts so the range of each gene remains the same after the
transformation.
}
\examples{
data("geneLengthGRCh38")
countMat<-sapply(vector("numeric",length = 100),function(x){
     c(MASS::rnegbin(10,mu = 50,theta = 5),
       MASS::rnegbin(10,mu = 10,theta = 5))
}) |> t(); countMat<-log2(countMat+1)
heatmap.DM(countMat)
#warning because of the small matrix
correctedMat<-oobFastMNN(countMat,batch = c(rep(1,10),rep(2,10)), k=5)
heatmap.DM(correctedMat)
}
