% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pca2d}
\alias{pca2d}
\title{Plot a PCA in 2D}
\usage{
pca2d(
  pca,
  comp = 1:2,
  colorBy = NULL,
  plotVars = FALSE,
  pointSize = 2,
  plotText = FALSE,
  ratioPerPercentVar = FALSE,
  main = NULL,
  ellipse = FALSE,
  colorScale = NULL,
  returnGraph = FALSE,
  outlierLabel = FALSE,
  outlierLabelThres = NULL,
  outlierLabelSize = 3,
  customRatio = NULL,
  ...
)
}
\arguments{
\item{pca}{PCA object created by `PCA` function.}

\item{comp}{A vector of 2 integer. Principal Component to be plotted.}

\item{colorBy}{A vector of factor or numeric same size as number of samples in PCA. Color each point on the projection by its corresponding value.}

\item{plotVars}{Logical. Plot variable loadings instead of coordinates of samples on PCs.}

\item{pointSize}{Single numeric. Size of points.}

\item{plotText}{Logical. Plot sample/variable name instead of point.}

\item{ratioPerPercentVar}{Logical. Ratio of the plot is corresponding to the percentage of explained variation from each PC.}

\item{main}{Single character. Main title of the graph.}

\item{ellipse}{Logical. Add stat_ellipse to the projection (eventually several ellipses if colorBy is a factor).}

\item{colorScale}{A vector of colors. Equal to the desired number of breaks for continuous values or number of levels for factors.}

\item{returnGraph}{Logical. Return the graph as a ggplot object instead of printing it.}

\item{outlierLabel}{Logical. Add a `ggrepel` label for point in low-density area of the plot.}

\item{outlierLabelThres}{Numeric. Density value threshold below that the labelled is displayed if `outlierLabel=TRUE`.}

\item{outlierLabelSize}{Numeric. Size of outlier labels.}

\item{customRatio}{Single numeric. Custom ratio for the plot.}

\item{...}{Arguments passed to `proj2d`}
}
\value{
Plot in the current graphical device or a ggplot object if `returnGraph=TRUE`.
}
\description{
Plot a PCA in 2D
}
\examples{
data(iris)
pca<-PCA(iris[,1:4],transpose = FALSE,scale = TRUE,center = TRUE)
pca2d(pca)
pca2d(pca,comp = c(1,3))

pca2d(pca,colorBy = iris$Species,ratioPerPercentVar = TRUE)
pca2d(pca,colorBy = iris$Species,colorScale=c("blue","white","red"))

pca2d(pca,plotVars = TRUE,plotText = TRUE,pointSize = 5)

pca2d(pca,outlierLabel = TRUE,colorBy = iris$Species)
pca2d(pca,outlierLabel = TRUE,colorBy = iris$Species,ellipse = TRUE,outlierLabelThres = .2)

pca2d(pca,outlierLabel = TRUE,colorBy = iris$Species,ellipse = TRUE,outlierLabelThres = .2,returnGraph = TRUE)+
	theme_dark()
}
