% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReductions.R
\name{pcaAddSamples}
\alias{pcaAddSamples}
\title{Add new samples to an existing PCA.}
\usage{
pcaAddSamples(pca, newSamplesMatrix, transpose = TRUE, returnPCA = TRUE)
}
\arguments{
\item{pca}{The existing PCA as an object returned by \code{PCA} or \code{fastPCA}.}

\item{newSamplesMatrix}{The matrix of numeric of new samples. It must have
the same features than the original PCA.}

\item{transpose}{Logical. If \code{transpose}, samples are columns and features
are rows.}

\item{returnPCA}{Logical. Return the updated PCA. If false, just returned the
coordinates of new samples on PCs.}
}
\value{
A PCA list if \code{returnPCA} or a matrix of coordinates of samples on
PCs.
}
\description{
Add new samples to an existing PCA.
}
\examples{
data("iris")
iris1<-iris[1:75,]
iris2<-iris[76:150,]
pca<-PCA(iris1[,seq_len(4)],transpose = FALSE,scale = TRUE,center = TRUE)
pca2d(pca,colorBy = iris1$Species)
pcaUpdated<-pcaAddSamples(pca,iris2[,seq_len(4)],transpose = FALSE)
pca2d(pcaUpdated,colorBy = iris$Species)
}
