% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReductions.R
\name{PCR}
\alias{PCR}
\title{Principal Component Regression}
\usage{
PCR(pca, annotationDF, nComponent = 10)
}
\arguments{
\item{pca}{A PCA as an object returned by \code{PCA} or \code{fastPCA}.}

\item{annotationDF}{A dataframe of feature (numeric or factor) with rows as samples. Must have the same number of samples than the PCA.}

\item{nComponent}{Integer, number of PC used in the regression.}
}
\value{
A dataframe with the PC, annotation and the corresponding R squared.
}
\description{
PCR is a way to link experimental variable to principal components.
}
\examples{
data("bulkLogCounts")
data("sampleAnnot")

pca<-fastPCA(bulkLogCounts,nPC=10)
PCR(pca,annotationDF=sampleAnnot[,c("culture_media","line","passage")])
}
