% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotExpr}
\alias{plotExpr}
\title{Plot the expression of one or several genes}
\usage{
plotExpr(
  expr,
  group = NULL,
  log10Plus1yScale = NULL,
  violin = TRUE,
  boxplot = TRUE,
  dotplot = FALSE,
  violinArgs = list(),
  boxplotArgs = list(),
  dotplotArgs = list(),
  colorScale = mostDistantColor,
  legendTitle = "group",
  dodge.width = 0.9,
  returnGraph = FALSE
)
}
\arguments{
\item{expr}{Numeric 2D matrix. Each row is gene and each column a sample. Row
must be named by genes.}

\item{group}{Factor vector, same length as number of column in expr.
Experimental group attributed to each sample.}

\item{log10Plus1yScale}{Logical or NULL. Use a \code{log10(x+1)} scale. By default
\code{FALSE} if one gene and \code{TRUE} if several.}

\item{violin}{Logical. Plot violin plots.}

\item{boxplot}{Logical. Plot violin plots.}

\item{dotplot}{Logical. Plot dot plots.}

\item{violinArgs}{List. Additional arguments given to \code{geom_violin}.}

\item{boxplotArgs}{List. Additional arguments given to \code{geom_boxplot}.}

\item{dotplotArgs}{List. Additional arguments given to \code{geom_beeswarm}.}

\item{colorScale}{A list of color. Must be the same length as number of
levels in group.}

\item{legendTitle}{Character. Displayed in legend title.}

\item{dodge.width}{Numeric. Width of individual distribution element
(violin/boxplot/dotplot).}

\item{returnGraph}{Logical. Print the ggplot object or return it.}
}
\value{
Plot in the current graphical device or a ggplot object if
\code{returnGraph=TRUE}.
}
\description{
Plot the expression of one or several genes
}
\examples{
exprMat<-rbind(
    MASS::rnegbin(30,theta = 10,mu = 10),
    MASS::rnegbin(30,theta = 1,mu = 10),
    MASS::rnegbin(30,theta = 1,mu = 1000)
)
rownames(exprMat)<-c("gene1","gene2","gene3")

group=c(rep("A",10),rep("B",10),rep("C",10))

plotExpr(exprMat)
plotExpr(exprMat["gene1",])
plotExpr(exprMat["gene1",],group = group)
plotExpr(exprMat,group = group)

plotExpr(exprMat,group = group,log10Plus1yScale = FALSE)
plotExpr(exprMat,group = group,dodge.width = .5,legendTitle = "Letter")
plotExpr(exprMat,boxplot = FALSE,violin = FALSE,dotplot = TRUE)
plotExpr(exprMat,group = group,colorScale = c("blue","white","red"))
plotExpr(exprMat,group = group,returnGraph = TRUE)+
    geom_point()

plotExpr(exprMat,group = group,violinArgs = list(scale="area"))
}
