% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPalette}
\alias{plotPalette}
\title{Plot colors}
\usage{
plotPalette(colorScale, continuousStep = NULL)
}
\arguments{
\item{colorScale}{Character vector, Colors to plot.}

\item{continuousStep}{Number of plotted color with intermediate between the
given colors.}
}
\value{
NULL, draw in current graphic device.
}
\description{
Plot colors
}
\examples{
plotPalette(c("#001D6E","white","#E6001F"))
plotPalette(c("#001D6E","white","#E6001F"),continuousStep = 50)
}
