% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pointdensity.nrd}
\alias{pointdensity.nrd}
\title{Compute density value for the point of a a 2D-space}
\usage{
pointdensity.nrd(mat, eps = 1)
}
\arguments{
\item{mat}{numeric matrix of point coordinates. Each row is a point, 1st
column = x, 2nd column=y.}

\item{eps}{radius of the eps-neighborhood, i.e., bandwidth of the uniform
kernel).}
}
\value{
A vector of numeric value of the length equal to the number of points
representing the density
}
\description{
Compute density value for the point of a a 2D-space
}
\examples{
coord<-matrix(rnorm(2000),ncol = 2)
pointdensity.nrd(coord)
}
