% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{proj3d}
\alias{proj3d}
\title{Plot a 3D projection.}
\usage{
proj3d(
  coord,
  axis = 1:3,
  colorBy = NULL,
  pointSize = 5,
  plotText = FALSE,
  colorScale = NULL,
  na.color = "grey50",
  na.bg = TRUE,
  alpha = 1,
  ...
)
}
\arguments{
\item{coord}{Matrix or dataframe containing at least three vectors of numeric corresponding to the x/y/z coordinates.}

\item{axis}{A vector of 3 integer. The column index of \code{coord} containing respectively the x and y coordinates.}

\item{colorBy}{A vector of factor or numeric same size as number of samples in PCA. Color each point on the projection by its corresponding value.}

\item{pointSize}{Single numeric. Size of points.}

\item{plotText}{Logical. Plot sample instead of point.}

\item{colorScale}{A vector of colors. Equal to the desired number of breaks for continuous values or number of levels for factors.}

\item{na.color}{Character. Color of NA values if \code{ColorBy} is provided.}

\item{na.bg}{Logical. Put the NA points behind the others in term of layers.}

\item{alpha}{Single numeric (from 0 to 1). Opacity of the points.}

\item{...}{Other arguments passed to \code{plot3d}.}
}
\value{
Plot the projection in a new rgl window.
}
\description{
Plot a 3D projection.
}
\examples{
data(iris)

proj3d(iris,pointSize = .05)
proj3d(iris,axis = c(1,3,4),pointSize = .05)


proj3d(iris,colorBy = iris$Species,pointSize = .05)

proj3d(iris,colorBy = iris$Species,colorScale = c("blue","white","red"),pointSize = .05)
}
