% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrix.R
\name{qauroc}
\alias{qauroc}
\title{Quick approximation of area under ROC curve}
\usage{
qauroc(score, boolVect)
}
\arguments{
\item{score}{a vector of numeric representing the measure of a feature in a set of samples.}

\item{boolVect}{a vector of logical, same size as score. Is the sample in the target group?}

\item{n1}{a single integer, number of observation not in the target group.}

\item{n2}{a single integer, number of observation in the target group.}
}
\value{
A numeric value. Close to 1 = perfect marker, around 0.5 = as good as random values, close to 0 = perfect anti-marker.
}
\description{
By Miron Kursa https://mbq.me
}
\examples{
data(iris)
qauroc(iris$Sepal.Length,iris$Species=="virginica", sum(!iris$Species=="virginica"),sum(iris$Species=="virginica"))
}
