% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot.R
\name{qplotAutoX}
\alias{qplotAutoX}
\title{Quick plot of values with x values corresponding to index.}
\usage{
qplotAutoX(x, returnGraph = FALSE, geom = "point", ...)
}
\arguments{
\item{x}{Vector of numeric.}

\item{returnGraph}{Logical. Print the ggplot object or return it.}

\item{geom}{Geom to draw. Defaults to "point"}

\item{...}{Parameters passed to \code{oobqplot}}
}
\value{
Plot in the current graphical device or a ggplot object if \code{returnGraph=TRUE}.
}
\description{
Quick plot of values with x values corresponding to index.
}
\examples{
qplotAutoX(seq_len(5))
}
