% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{qplotBarplot}
\alias{qplotBarplot}
\title{Quick simple barplot with heights provided in ggplot format.}
\usage{
qplotBarplot(y, returnGraph = FALSE)
}
\arguments{
\item{y}{Vector of numeric.}

\item{returnGraph}{Logical. Print the ggplot object or return it.}
}
\value{
Plot in the current graphical device or a ggplot object if \code{returnGraph=TRUE}.
}
\description{
Quick simple barplot with heights provided in ggplot format.
}
\examples{
qplotBarplot(1:5)
y<-1:5
names(y)<-intToUtf8(65:69,multiple = TRUE)
qplotBarplot(y)
}
