% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{qplotDensity}
\alias{qplotDensity}
\title{Plot a density-based distribution of a feature.}
\usage{
qplotDensity(x, returnGraph = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{returnGraph}{Logical. Logical. Return the graph as a ggplot object instead of printing it.}

\item{...}{Parameters passed to \code{qplot}}
}
\value{
Plot in the current graphical device or a ggplot object if \code{returnGraph=TRUE}.
}
\description{
Plot a density-based distribution of a feature.
}
\examples{
qplotDensity(rnorm(10000))
}
