% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrix.R
\name{quickSCnorm}
\alias{quickSCnorm}
\title{Quick single cell normalization (scran method).}
\usage{
quickSCnorm(rawCounts, returnLog = TRUE, sizeFactors = NULL, ...)
}
\arguments{
\item{rawCounts}{RNA-Seq raw counts with rows as genes.}

\item{returnLog}{Logical. Return log counts.}

\item{sizeFactors}{NULL or a vector of numeric containing precomputed size factor, same size as number of cells in `rawCounts`.}

\item{...}{Other parameter passed to `computeSumFactors`.}
}
\value{
A normalized count table.
}
\description{
Quick single cell normalization (scran method).
}
\examples{
data("geneLengthGRCh38")
library(MASS)
countMat<-t(sapply(vector("numeric",length = length(geneLengthGRCh38)),function(x){
	rnegbin(10,theta = abs(rnorm(1,mean = 10,sd = 20)),mu = abs(rnorm(1,mean = 10,sd = 20)))
}));rownames(countMat)<-names(geneLengthGRCh38)
quickSCnorm(countMat, returnLog=FALSE)
quickSCnorm(countMat, returnLog=TRUE)
}
