% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read.vectorList}
\alias{read.vectorList}
\title{Load a list from a text file.}
\usage{
read.vectorList(fileName, sep = "\\t")
}
\arguments{
\item{fileName}{Path to the file to load}

\item{sep}{Separator ("\\t"=tab-separated values).}
}
\value{
A list.
}
\description{
Currently works for file saved with parameters list.names=TRUE
and vector.names=FALSE from write.vectorList.
}
\examples{
tf <- tempfile(fileext = ".tsv")
write.vectorList(list(1:10,letters[1:10]),tf)
read.vectorList(tf)
}
\seealso{
write.vectorList
}
