% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDataframe.R
\name{reScale}
\alias{reScale}
\title{Put rows of a matrix on the same range than another.}
\usage{
reScale(matToAdjust, matGoodRange.)
}
\arguments{
\item{matToAdjust}{Matrix of numeric where the range of each row has to be adjusted.}

\item{matGoodRange.}{Matrix of numeric, same dimension as \code{matToAdjust}.}
}
\value{
Return \code{matToAdjust} but the each row has now same minimum and maximum to its corresponding row in matGoodRange.
}
\description{
Put rows of a matrix on the same range than another.
}
\examples{
m1<-matrix(rnorm(100),ncol=50)
qplotDensity(m1[1,])
m2<-matrix(rnorm(100,20),ncol=50)
qplotDensity(m2[1,])
m1Rescale<-reScale(m1,m2)
qplotDensity(m1Rescale[1,])
}
