% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{richUpsetStats}
\alias{richUpsetStats}
\title{Compute enrichment statistics for set intersection.}
\usage{
richUpsetStats(featurePerGroupList, universe = NULL)
}
\arguments{
\item{featurePerGroupList}{A list of sets (vector of character)}

\item{universe}{NULL or vector of Character. The entire list of features (Universal set)}
}
\value{
A list that contains:
\enumerate{
\item A data frame (\emph{enrichStats}) containing the enrichment statistics for each intersection of sets.
Row names are corresponding to the intersection code (e.g. 101 for the intersection of the first and third set).
The columns are:
\itemize{
\item \code{observed}: observed number of elements in intersection
\item \code{expected}: expected number of elements in intersection
\item \code{OEdeviation}: Effect size of the association \code{(observed - expected) / sqrt(universeSize)}
\item \code{pval}: pvalue (upper tail) where NULL hypotheses is that observed intersection is from the distribution of intersection size with \emph{p = nExpected / nUniverse}.
\item \code{padj}: p-value adjusted for multiple testing
\item \code{log10padj}: -log10(padj), with a floor at 384
\item \code{combDegree}: Number of sets in the intersection
\item \code{combsize}: Number of elements in the intersection
}
\item A combination matrix (\emph{upsetMatrix}), can be used as input to \link[ComplexHeatmap:UpSet]{ComplexHeatmap::UpSet}
\item \emph{universeSize}: Total number of element in the universe.
\item \emph{setsize}: Number of elements in each set.
}
}
\description{
Compute enrichment statistics for set intersection.
}
\examples{
lt = list(set1 = sample(letters, 5),
                    set2 = sample(letters, 10),
                    set3 = sample(letters, 15)
                    )
lt$set4 <- unique(c(lt$set1,lt$set2))

richUpsetStats(lt, universe = letters)

}
\seealso{
\code{\link[=richUpset]{richUpset()}}
}
