% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rn<-}
\alias{rn<-}
\title{rownames alias (setter)}
\usage{
rn(x) <- value
}
\arguments{
\item{x}{A matrix-like object.}

\item{value}{Either NULL or a character vector equal of length equal to the
appropriate dimension.}
}
\value{
The modified object.
}
\description{
rownames alias (setter)
}
\examples{
data(iris)
rn(iris)<-paste0("f",nrow(iris) |> seq_len())
rn(iris)
}
