% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rn}
\alias{rn}
\title{rownames alias (getter)}
\usage{
rn(x, do.NULL = TRUE, prefix = "row")
}
\arguments{
\item{x}{A matrix-like object.}

\item{do.NULL}{logical. If \code{FALSE} and names are \code{NULL}, names are created.}

\item{prefix}{for created names}
}
\value{
A character vector.
}
\description{
rownames alias (getter)
}
\examples{
data(iris)
rn(iris)
}
