% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrix.R
\name{RPKM}
\alias{RPKM}
\title{Reads Per Kilobase per Million (RPKM) normalization for full-length
transcript, short read sequencing.}
\usage{
RPKM(data, gene.length, sce_assay = 1)
}
\arguments{
\item{data}{A raw count table with rows as genes.
Can also be a \code{SummarizedExperiment} or \code{SingleCellExperiment}  object.}

\item{gene.length}{A vector of numeric corresponding to gene length in base
pair. Must be named by genes.}

\item{sce_assay}{Integer or character, if \code{data} is a
\code{SummarizedExperiment} related object, the assay name to use.}
}
\value{
A normalized count table of RPKM.
If \code{data} is a \code{SummarizedExperiment} object, the function will add the
normalized count to the \code{SummarizedExperiment} object as an assay named
\code{rpkm}.
}
\description{
Reads Per Kilobase per Million (RPKM) normalization for full-length
transcript, short read sequencing.
}
\examples{
data("geneLengthGRCh38")
countMat <-
    sapply(vector("numeric", length = length(geneLengthGRCh38)), function(x){
        MASS::rnegbin(10, theta = abs(rnorm(1, mean = 10, sd = 20)),
                     mu = abs(rnorm(1, mean = 10, sd = 20)))
    }) |> t()
rownames(countMat)<-names(geneLengthGRCh38)
chead(RPKM(countMat,geneLengthGRCh38))

sce <- SingleCellExperiment(assays = list(counts = countMat))
assay(RPKM(sce,geneLengthGRCh38),"rpkm") |> chead()
}
