% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oob-package.R
\docType{data}
\name{sampleAnnot}
\alias{sampleAnnot}
\title{sampleAnnot}
\format{
A data.frame containing 77 observations for 18 features.
}
\description{
Sample Annotation sheet (often referred as colData) for 77 bulk
RNA-Seq samples used in Kilens, Meistermann et al. 2018.
}
\references{
\url{https://doi.org/10.1038/s41467-017-02107-w}
}
\author{
Dimitri Meistermann
}
\keyword{data}
