% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setRandState}
\alias{setRandState}
\title{Set the precise random seed state}
\usage{
setRandState(state)
}
\arguments{
\item{state}{Object saved by getRandState}
}
\description{
Set the precise random seed state
}
