% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{simpleAbundanceAnalysis}
\alias{simpleAbundanceAnalysis}
\title{Very simple over abundance analysis via logistic regression}
\usage{
simpleAbundanceAnalysis(varToExplain, df, cols, removeIntercepts = TRUE)
}
\arguments{
\item{df}{A DataFrame, same number of rows as length of qualVar.}

\item{cols}{Vector of character. Column names of df to be used in the regression model.}

\item{removeIntercepts}{Should the intercept be removed from the model? Useful if odds ratio / pvalue of all levels of a factor imncluded in cols are wanted.}

\item{qualVar}{A vector or factor or character.}
}
\value{
A dataframe containing 2 column per level of qualVar (Odds Ratio and pvalues).
}
\description{
Very simple over abundance analysis via logistic regression
}
\examples{
data(sampleAnnot)
simpleAbundanceAnalysis(sampleAnnot$culture_media, sampleAnnot,  c("line","run"))

}
