% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sortColorByDistance}
\alias{sortColorByDistance}
\title{Sort a vector of color by their similarity}
\usage{
sortColorByDistance(colorVector, byDissimilarity = FALSE)
}
\arguments{
\item{colorVector}{Vector of colors.}

\item{byDissimilarity}{Order by dissimilarity instead of similarity}
}
\value{
A vector of colors, sorted.
}
\description{
Sort a vector of color by their similarity
}
\details{
The function proceeds as follows:
\enumerate{
\item Convert the color vector to the Lab color space.
\item Compute the distance between each color.
\item Order the colors using a hierarchical clustering.
}
}
\examples{
colors <- c("#000066","#660000","#006600","#0000FF","#FF0000","#00FF00")
sortColorByDistance(colors) |> plotPalette()
colors <- c("#000066","#0000FF","#660000","#FF0000","#006600","#00FF00")
sortColorByDistance(colors, byDissimilarity=TRUE) |> plotPalette()
}
