% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDataframe.R
\name{subSampleColumnPerGroup}
\alias{subSampleColumnPerGroup}
\title{Drawn samples from each population in a matrix where columns are samples.}
\usage{
subSampleColumnPerGroup(mat, groupVector, n = NULL)
}
\arguments{
\item{mat}{A matrix where each column a samples with features as rows.}

\item{groupVector}{A vector of character or factor, same length as number of column in \code{mat}. Describe the population of each sample.}

\item{n}{Integer or NULL. The number of samples to be draw in each population. If NULL, equal to the smallest population.}
}
\value{
A matrix of drawn samples.
}
\description{
Drawn samples from each population in a matrix where columns are samples.
}
\examples{
mat<-matrix(rnorm(100),ncol=10,dimnames = list(paste0("feature",1:10),paste0("sample",1:10)))
groups<-rep(c("A","B"),each=5)
subSampleColumnPerGroup(mat,groups,n=3)
}
