% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDataframe.R
\name{supprNAnames}
\alias{supprNAnames}
\title{Delete row/column in a matrix/df with NA names}
\usage{
supprNAnames(x, side = 1)
}
\arguments{
\item{x}{A dataframe or matrix}

\item{side}{1:row names are treated, 2: column names.}
}
\value{
A matrix or dataframe. Same format as x.
}
\description{
Delete row/column in a matrix/df with NA names
}
\examples{
x<-matrix(1:25,ncol=5)
rownames(x)<-c(letters[1:4],NA)
supprNAnames(x)
}
