% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{takefirst}
\alias{takefirst}
\title{Take first element of multiple values in a vector}
\usage{
takefirst(x, returnIndex = FALSE)
}
\arguments{
\item{x}{Vector.}

\item{returnIndex}{Logical. Should the index of first elements or vector of
first elements.}
}
\value{
Vector of first elements or numeric vector of index.
}
\description{
Similar to unique but conserve vector names or return index
where you can find each first value of multiple element.
}
\examples{
a<-c(1,2,3,3,4,3)
names(a)<-c("a","b","c","d","e","f")
takefirst(a)
takefirst(a,returnIndex = TRUE)
}
