% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReductions.R
\name{TRIMAP}
\alias{TRIMAP}
\title{TriMap dimension reduction.}
\usage{
TRIMAP(
  data,
  n_dims = 2,
  transpose = TRUE,
  n_inliers = 10,
  n_outliers = 5,
  apply_pca = TRUE,
  n_iters = 400,
  knn_tuple = NULL,
  sce_assay = "logcounts"
)
}
\arguments{
\item{data}{A matrix of numeric (in the RNA-Seq context, log counts).
Can also be a SingleCellExperiment object.}

\item{n_dims}{Integer. Dimensions of the embedded space.}

\item{transpose}{Logical. If \code{transpose}, samples are columns and features
are rows.}

\item{n_inliers}{Number of nearest neighbors for forming the nearest neighbor
triplets.}

\item{n_outliers}{Number of outliers for forming the nearest neighbor
triplets.}

\item{apply_pca}{Reduce the number of dimensions of the data to 100 if
necessary before applying the nearest-neighbor search.}

\item{n_iters}{Number of iterations.}

\item{knn_tuple}{Use the precomputed nearest-neighbors information in form of
a tuple (knn_nbrs, knn_distances).}

\item{sce_assay}{Integer or character,
if \code{data} is a \code{SingleCellExperiment} object, the assay name to use.}
}
\value{
A matrix of coordinates with samples as rows.

If \code{data} is a \code{SingleCellExperiment} returns a \code{SingleCellExperiment}
with the TriMap computed in the \code{reducedDims} slot at the index \code{TriMap}.
}
\description{
TriMap dimension reduction.
}
\examples{
data(iris)
irisTrimap<-TRIMAP(iris[,seq_len(4)],transpose=FALSE)
proj2d(irisTrimap,colorBy = iris$Species)

irisSCE <- SingleCellExperiment::SingleCellExperiment(
        assays = list(counts = t(iris[, -5])),
        colData = iris[5]
)
irisSCE<-TRIMAP(irisSCE)
}
