% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReductions.R
\name{UMAP}
\alias{UMAP}
\title{UMAP projection}
\usage{
UMAP(
  data,
  nDimPCA = NULL,
  transpose = TRUE,
  n_neighbors = 20,
  n_components = 2,
  min_dist = 0.01,
  init = "laplacian",
  metric = "euclidean",
  ret_model = FALSE,
  ret_nn = FALSE,
  sce_assay = 1,
  ...
)
}
\arguments{
\item{data}{A matrix of numeric (in the RNA-Seq context, log counts).
Can also be a SingleCellExperiment object.}

\item{nDimPCA}{Integer. If not NULL compute a PCA first and take n first
principal components.}

\item{transpose}{Logical. If \code{transpose}, samples are columns and features
are rows.}

\item{n_neighbors}{The size of local neighborhood (in terms of number of
neighboring sample points) used for manifold approximation. Larger values
result in more global views of the manifold, while smaller values result in
more local data being preserved. In general values should be in the range 2
to 100.}

\item{n_components}{Integer. Dimensions of the embedded space.}

\item{min_dist}{The effective minimum distance between embedded points.
Smaller values will result in a more clustered/clumped embedding where
nearby points on the manifold are drawn closer together, while larger
values will result on a more even dispersal of points. The value should be
set relative to the spread value, which determines the scale at which
embedded points will be spread out.}

\item{init}{Type of initialization for the coordinates (see \code{uwot} for more
details).}

\item{metric}{Type of distance metric to use to find nearest neighbors (see
\code{uwot} for more details).}

\item{ret_model}{If TRUE, then return extra data that can be used to add new
data to an existing embedding via umap_transform. The embedded coordinates
are returned as the list item embedding. If FALSE, just return the
coordinates. This parameter can be used in conjunction with ret_nn and
ret_extra. Note that some settings are incompatible with the production of
a UMAP model: external neighbor data (passed via a list to nn_method), and
factor columns that were included via the metric parameter. In the latter
case, the model produced is based only on the numeric data. A
transformation using new data is possible, but the factor columns in the
new data are ignored}

\item{ret_nn}{If TRUE, then in addition to the embedding, also return nearest
neighbor data that can be used as input to nn_method to avoid the overhead
of repeatedly calculating the nearest neighbors when manipulating unrelated
parameters (e.g. min_dist, n_epochs, init). See the "Value" section for the
names of the list items. If FALSE, just return the coordinates. Note that
the nearest neighbors could be sensitive to data scaling, so be wary of
reusing nearest neighbor data if modifying the scale parameter. This
parameter can be used in conjunction with ret_model and ret_extra.}

\item{sce_assay}{Integer or character,
if \code{data} is a \code{SingleCellExperiment} object, the assay name to use.}

\item{...}{Other parameters passed to uwot.}
}
\value{
A matrix of coordinates with samples as rows, or:
\itemize{
\item if ret_model = TRUE (or ret_extra contains "model"),
returns a list containing extra information that can be used to add new
data to an existing embedding via umap_transform.
In this case, the coordinates are available in the list item embedding.
NOTE: The contents of the model list should not be considered stable or
part of the public API, and are purposely left undocumented.
\item if ret_nn = TRUE (or ret_extra contains "nn"),
returns the nearest neighbor data as a list called nn.
This contains one list for each metric calculated,
itself containing a matrix idx with the integer ids of the neighbors;
and a matrix dist with the distances. The nn list (or a sub-list)
can be used as input to the nn_method parameter.
\item if data is a SingleCellExperiment object, returns a SingleCellExperiment
object with the UMAP coordinates in the \code{reducedDims} slot, and the eventual
complete UMAP model and nearest neighbor in the metadata slot at the index
\code{UMAP}.
}
}
\description{
UMAP projection
}
\examples{
data(iris)
irisUMAP<-UMAP(iris[,seq_len(4)],transpose = FALSE)
proj2d(irisUMAP,colorBy = iris$Species)
irisUMAP<-UMAP(rowScale(iris[,seq_len(4)],center = TRUE,scaled = TRUE),
    transpose = FALSE,n_neighbors = nrow(iris),
    ret_nn = TRUE, ret_model = TRUE)
proj2d(irisUMAP$embedding,colorBy = iris$Species,
    nnMatrix = irisUMAP$nn$euclidean$idx[,c(1, 2, 3)],fixedCoord = TRUE)

irisSCE <- SingleCellExperiment::SingleCellExperiment(
        assays = list(counts = t(iris[, -5])),
        colData = iris[5]
)
irisSCE<-UMAP(irisSCE)
}
