% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uncenter}
\alias{uncenter}
\title{Transform vector to have no negative value (new minimum is 0)}
\usage{
uncenter(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector
}
\description{
Transform vector to have no negative value (new minimum is 0)
}
\examples{
uncenter(-5:5)
}
