% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{VectorListToFactor}
\alias{VectorListToFactor}
\title{Convert a list to a named factor vector}
\usage{
VectorListToFactor(listOfVector)
}
\arguments{
\item{listOfVector}{A named list. Each element must contain a character
vector.}
}
\value{
A named factor vector.
}
\description{
Convert a list to a named factor vector
}
\examples{
VectorListToFactor(list(a=c("x1","x2"),b=c("x3","x4"),c=c("x5","x6","x7")))

}
\seealso{
factorToVectorList
}
