% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{viewKEGG}
\alias{viewKEGG}
\title{Wrapper for pathview, with gene symbols as input}
\usage{
viewKEGG(
  x,
  pathway,
  corrIdGenes = NULL,
  species = "Human",
  speciesData = NULL,
  directory = getwd(),
  ...
)
}
\arguments{
\item{x}{either vector (single sample) or a matrix-like data (multiple sample). Vector should be numeric with gene IDs as names or it may also be character of gene IDs. Character vector is treated as discrete or count data. Matrix-like data structure has genes as rows and samples as columns. Row names should be gene IDs. Here gene ID is a generic concepts, including multiple types of gene, transcript and protein uniquely mappable to KEGG gene IDs. KEGG ortholog IDs are also treated as gene IDs as to handle metagenomic data. Check details for mappable ID types. Default gene.data=NULL.}

\item{pathway}{character vector, the KEGG pathway ID(s), usually 5 digit, may also include the 3 letter KEGG species code.}

\item{corrIdGenes}{Dataframe of gene ID correspondence where each column is a gene ID type. If not NULL `species` and `speciesData` arguments wont be used.}

\item{species}{Character. Shortname of the species as described in `data("bods")`.}

\item{speciesData}{object returned by `getSpeciesData`. If not NULL `species` argument wont be used.}

\item{...}{Other arguments passed to `pathview`.}

\item{kegg.dir}{character, the directory of KEGG pathway data file (.xml) and image file (.png). Users may supply their own data files in the same format and naming convention of KEGG's (species code + pathway id, e.g. hsa04110.xml, hsa04110.png etc) in this directory. Default kegg.dir="." (current working directory).}
}
\value{
File congaing pathway scheme and projection of x values on it.
}
\description{
Wrapper for pathview, with gene symbols as input
}
