% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{volcanoPlot.DESeq2}
\alias{volcanoPlot.DESeq2}
\title{Volcano plot with additional annotation for interpreting DE genes.}
\usage{
volcanoPlot.DESeq2(
  DEresult,
  formula,
  downLevel,
  upLevel,
  condColumn,
  padjThreshold = 0.05,
  LFCthreshold = 1,
  topGene = 30
)
}
\arguments{
\item{DEresult}{Dataframe that contains at least those columns:
\itemize{
\item padj (adjusted p-value)
\item isDE (a character vector equal to "NONE" if the gene is not DE, "DOWNREG" or "UPREG" if DE).
\item log2FoldChange
Row must be named by genes.
}}

\item{formula}{Character. Design formula given to DESeq2.}

\item{downLevel}{Character. Condition considered as the reference. If a gene is more expressed in this condition, LFC < 0.}

\item{upLevel}{Character. Condition considered as the target group. If a gene is more expressed in this condition, LFC > 0.}

\item{condColumn}{Character. Name of the experimental variable that have been used for differential expression.}

\item{padjThreshold}{Numeric. Significance threshold of the adjusted p-value.}

\item{LFCthreshold}{Numeric. Significance threshold of the Log2 Fold-Change.}

\item{topGene}{Integer. Number of gene name to be shown on the plot. Genes names are plotted from the most significant.}
}
\value{
Plot in the current graphical device.
}
\description{
Volcano plot with additional annotation for interpreting DE genes.
}
\examples{
data("DEgenesPrime_Naive")
volcanoPlot.DESeq2(DEgenesPrime_Naive,formula = "~culture_media+Run",condColumn = "culture_media",downLevel = "KSR+FGF2",upLevel = "T2iLGO")
}
