% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{whichTop}
\alias{whichTop}
\title{Return ordered index of element with top n value.}
\usage{
whichTop(x, top = 5, decreasing = TRUE)
}
\arguments{
\item{x}{Numeric vector.}

\item{top}{Integer. Number of element to be returned.}

\item{decreasing}{Logical, return top element by decreasing or increasing
order.}
}
\value{
A vector of integer.
}
\description{
Return ordered index of element with top n value.
}
\examples{
x<-c(1,5,6,10,5.2,3,8)
whichTop(x)
whichTop(x,top=3)
whichTop(x,decreasing=FALSE)
}
