% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixDataframe.R
\name{aggregMatPerVector}
\alias{aggregMatPerVector}
\title{Aggregate columns or rows of a matrix by a vector of factor}
\usage{
aggregMatPerVector(x, by, FUN = mean, byRow = NULL)
}
\arguments{
\item{x}{A matrix of numeric.}

\item{by}{A vector of factor. Same size as number of row in `x` if  `byRow`, otherwise same size as number of columns.}

\item{FUN}{a function to compute the summary statistics which can be applied to all data subsets.}

\item{byRow}{Logical. Do the aggregation by row or by column. If NULL, determines automatically based on the size of `by`.}
}
\value{
A matrix of numeric. Aggregated dimension names are changed to match the levels of `by`.
}
\description{
Aggregate columns or rows of a matrix by a vector of factor
}
\examples{
data(iris)
aggregMatPerVector(iris[,1:4],iris$Species)
}
