% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projPlots.R
\name{barplotPercentVar}
\alias{barplotPercentVar}
\title{Scree plot of explained variance per PC from a PCA}
\usage{
barplotPercentVar(pca, nPC = length(pca$propExplVar), returnGraph = FALSE)
}
\arguments{
\item{pca}{PCA object created by \code{PCA} function.}

\item{nPC}{Integer. Number of principal component to be plotted.}

\item{returnGraph}{Logical. Print the ggplot object or return it.}
}
\value{
Plot in the current graphical device or a ggplot object if
\code{returnGraph=TRUE}.
}
\description{
Scree plot of explained variance per PC from a PCA
}
\examples{
data(iris)
pca<-PCA(iris[,seq_len(4)],transpose = FALSE,scale = TRUE,center = TRUE)
barplotPercentVar(pca)
barplotPercentVar(pca, nPC = 2)
}
