% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrix.R
\name{computeQCmetricSamples}
\alias{computeQCmetricSamples}
\title{Give QC metrics per sample}
\usage{
computeQCmetricSamples(x, uncenter = FALSE)
}
\arguments{
\item{x}{A matrix of numeric with samples as columns.}

\item{uncenter}{uncenter the matrix first so the minimum is 0.}
}
\value{
A data frame with each row as a sample and the following columns:
* `mean`: average expression in the sample
* `sd`: standard deviation
* `TotalGenEx`: number of expressed gene (count>0) in the sample
* `TotalCount`: sum of counts for the sample
}
\description{
Give QC metrics per sample
}
\examples{
data("geneLengthGRCh38")
library(MASS)
countMat<-t(sapply(vector("numeric",length = length(geneLengthGRCh38)),function(x){
	rnegbin(10,theta = abs(rnorm(1,mean = 10,sd = 20)),mu = abs(rnorm(1,mean = 10,sd = 20)))
}));rownames(countMat)<-names(geneLengthGRCh38)
colnames(countMat)<-letters[1:ncol(countMat)]
computeQCmetricSamples(countMat)
}
