% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrix.R
\name{computeQCmetricSamples}
\alias{computeQCmetricSamples}
\title{Give QC metrics per sample}
\usage{
computeQCmetricSamples(x, uncenter = FALSE)
}
\arguments{
\item{x}{A matrix of numeric with samples as columns.}

\item{uncenter}{uncenter the matrix first so the minimum is 0.}
}
\value{
A data frame with each row as a sample and the following columns:
\itemize{
\item \code{mean}: average expression in the sample
\item \code{sd}: standard deviation
\item \code{TotalGenEx}: number of expressed gene (count>0) in the sample
\item \code{TotalCount}: sum of counts for the sample
}
}
\description{
Give QC metrics per sample
}
\examples{
data("geneLengthGRCh38")
countMat<-t(sapply(vector("numeric",length = length(geneLengthGRCh38)),function(x){
    MASS::rnegbin(10,theta = abs(rnorm(1,mean = 10,sd = 20)),mu = abs(rnorm(1,mean = 10,sd = 20)))
}));rownames(countMat)<-names(geneLengthGRCh38)
colnames(countMat)<-letters[seq_len(ncol(countMat))]
computeQCmetricSamples(countMat)
}
