% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrix.R
\name{corGeneToOthers}
\alias{corGeneToOthers}
\title{Correlation from one gene to all others.}
\usage{
corGeneToOthers(gene, expression, corFun = cor, ...)
}
\arguments{
\item{gene}{A single charachter. The gene (or feature) name that will be used for correlating to all others.}

\item{expression}{A matrix of numeric with rows as features (in the RNA-Seq context, log counts).}

\item{corFun}{A function to compute a correlation between two feature.}

\item{...}{Parameters passed to \code{corFun}.}
}
\value{
A vector of numeric. Correlations values named by their corresponding gene.
}
\description{
Correlation from one gene to all others.
}
\examples{
data("bulkLogCounts")
corGeneToOthers("NANOG",bulkLogCounts)
}
