% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrix.R
\name{CPM}
\alias{CPM}
\title{Count Per Million normalization}
\usage{
CPM(data)
}
\arguments{
\item{data}{A raw count table with rows as genes}
}
\value{
A normalized count table.
}
\description{
Count Per Million normalization
}
\examples{
data("geneLengthGRCh38")
library(MASS)
countMat<-t(sapply(vector("numeric",length = length(geneLengthGRCh38)),function(x){
	rnegbin(10,theta = abs(rnorm(1,mean = 10,sd = 20)),mu = abs(rnorm(1,mean = 10,sd = 20)))
}));rownames(countMat)<-names(geneLengthGRCh38)
CPM(countMat)
}
