% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers_diffExpr.R
\name{extractFeatureMarkerData}
\alias{extractFeatureMarkerData}
\title{Extract a specific feature/metric (pval, logFC...) from a marker result dataframe}
\usage{
extractFeatureMarkerData(markerData, feature = "score")
}
\arguments{
\item{markerData}{A dataframe returned by \link{getMarkers}.}

\item{feature}{The name of the feature that has to be extracted.}
}
\value{
A matrix containing only the wanted feature where each column is a group.
}
\description{
Extract a specific feature/metric (pval, logFC...) from a marker result dataframe
}
\examples{
data("bulkLogCounts")
data("sampleAnnot")
markerData <- getMarkers(bulkLogCounts, sampleAnnot$culture_media)
extractFeatureMarkerData(markerData) |> chead()
}
