% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReductions.R
\name{fastPCA}
\alias{fastPCA}
\title{Faster Principal Component Analysis}
\usage{
fastPCA(
  data,
  transpose = TRUE,
  scale = FALSE,
  center = TRUE,
  nPC = min(ncol(data) - 1, nrow(data) - 1, 30),
  weight.by.var = TRUE,
  sce_assay = "logcounts",
  ...
)
}
\arguments{
\item{data}{A matrix of numeric (in the RNA-Seq context, log counts).
Can also be a SingleCellExperiment object.}

\item{transpose}{Logical. If \code{transpose}, samples are columns and features
are rows.}

\item{scale}{Logical. Divide features by their standard deviation.}

\item{center}{Logical. Subtract features by their average.}

\item{nPC}{Integer. Number of Principal Component to be computed.}

\item{weight.by.var}{Logical. Weight PC by variables. If TRUE return a
regular PCA.}

\item{sce_assay}{Integer or character,
if \code{data} is a \code{SingleCellExperiment} object, the assay name to use.}

\item{...}{Other parameters passed to \code{irlba}.}
}
\value{
\itemize{
\item sdev: the standard deviations of the principal components
(i.e., the square roots of the eigenvalues of the covariance/correlation
matrix, though the calculation is actually done with the singular
values of the data matrix).
\item rotation: the matrix of variable loadings
(i.e., a matrix whose columns contain the eigenvectors).
The function princomp returns this in the element loadings.
\item center, scale: the centering and scaling used, or FALSE.
\item x: the value of the rotated data (PC coordinates)
\item n.obs: Number of samples
\item propExplVar: proportion of explained variance by PCs on on variance of
computed axes.
\item transform: a list containing the scaling (sdeviations) and centering
factors (means) for each principal component.
\item isFastPCA: Logical. If TRUE, the PCA was computed using the \code{fastPCA}
}

If \code{data} is a \code{SingleCellExperiment}, returns a \code{SingleCellExperiment}
with the PCA computed in the \code{reducedDims} slot. The complete PCA object
is stored in the \code{metadata} slot at the index \code{PCA}.
}
\description{
Faster Principal Component Analysis
}
\examples{
data("bulkLogCounts")
pca<-fastPCA(bulkLogCounts,nPC=10)
pca2d(pca)

sce <- SingleCellExperiment(assays = list(counts = bulkLogCounts))
sce<-fastPCA(sce,nPC=10)
reducedDims(sce)
}
