% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrix.R
\name{getMarkers}
\alias{getMarkers}
\title{Compute a matrix of AUROC describing best gene marker per group of samples.}
\usage{
getMarkers(expressionMatrix, group, BPPARAM = NULL)
}
\arguments{
\item{expressionMatrix}{A matrix of numeric with rows as features (in the RNA-Seq context, log counts).}

\item{group}{A feature of factor/character, same length as number of sample. Describe group of each sample.}

\item{BPPARAM}{A BPPARAM object as return by `BiocParallel::bpparam()`. Use for multi-threading.}
}
\value{
A matrix of AUROC where each row is a gene and each column a level of `group`.
}
\description{
Compute a matrix of AUROC describing best gene marker per group of samples.
}
\examples{
data("bulkLogCounts")
data("sampleAnnot")
res <- getMarkers(bulkLogCounts,sampleAnnot$culture_media)
}
