% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrix.R
\name{getMostVariableGenes}
\alias{getMostVariableGenes}
\title{Compute over dispersion values for each gene.}
\usage{
getMostVariableGenes(counts, minCount = 0.01, plot = TRUE, returnPlot = FALSE)
}
\arguments{
\item{counts}{Normalized count table with genes as rows.}

\item{minCount}{Minimum average expression to not be filtered out.}

\item{plot}{Logical. Show the overdispersion plot.}

\item{returnPlot}{Logical, if `plot` return it as a ggplot object instead of printing it.}
}
\value{
A ggplot graph if `returnPlot`, otherwise a dataframe with the following columns:
- mu: average expression
- var: variance
- cv2: squared coefficient of variation. Used as a dispersion value.
- residuals: y-distance from teh regression. Can be used as an overdispersion value.
- residuals2: squared residuals
- fitted: theoretical dispersion for the gene average (y value of the curve).
}
\description{
Compute over dispersion values for each gene.
}
\examples{
data("bulkLogCounts")
normCount<-2^(bulkLogCounts-1)
dispData<-getMostVariableGenes(normCount,minCount=1)
}
