% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{GSDA.HeatmapAnnot}
\alias{GSDA.HeatmapAnnot}
\title{Displaying most neg and pos gene contribution on a GSDA heatmap.}
\usage{
GSDA.HeatmapAnnot(
  contributions,
  maxGeneContribAtOneSide = 3,
  width = unit(3, "cm"),
  fontsizeFactor = 400
)
}
\arguments{
\item{contributions}{A list named by pathway. Contains vector of gene contribution to activation scores, named by genes.}

\item{maxGeneContribAtOneSide}{Integer. Number of best pos/neg rank displayed on the annotation}

\item{width}{Numeric. Width of the annotation heatmap.}

\item{fontsizeFactor}{Numeric. Font-size of gene names.}
}
\value{
A HeatmapAnnotation object.
}
\description{
Displaying most neg and pos gene contribution on a GSDA heatmap.
}
