% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers_diffExpr.R
\name{multiLinearModel}
\alias{multiLinearModel}
\title{Test a linear model on each gene following an experimental design.}
\usage{
multiLinearModel(exprData, colData, contrast)
}
\arguments{
\item{exprData}{A matrix of numeric with rows as features (in the RNA-Seq context, log counts).}

\item{colData}{A dataframe of feature (numeric or factor) with rows as samples. Must have the same number of samples than exprData}

\item{contrast}{A vector of 3 character.
\enumerate{
\item Name of the experimental variable that have to be used for differential activation. Must be a column name of \code{colData}.
\item Condition considered as the reference.
\item Condition considered as the target group.
}}
}
\value{
A dataframe with the following columns:
\itemize{
\item baseMean: mean
\item log2FoldChange: Log(Log Fold Change)  between the two tested groups.
\item pval: an enrichment p-value
\item padj: a BH-adjusted p-value
}
}
\description{
Test a linear model on each gene following an experimental design.
}
\examples{
data("bulkLogCounts")
data("sampleAnnot")
res<-multiLinearModel(bulkLogCounts,colData = sampleAnnot,
    contrast = c("culture_media","T2iLGO","KSR+FGF2"))
}
