% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrix.R
\name{NMDS}
\alias{NMDS}
\title{Non-metric Multidimensional Scaling (dimension reduction)}
\usage{
NMDS(
  data,
  transpose = TRUE,
  scale = FALSE,
  center = FALSE,
  metric = dist,
  ndim = 2,
  maxit = 100
)
}
\arguments{
\item{data}{A matrix of numeric (in the RNA-Seq context, log counts).}

\item{transpose}{Logical. If `transpose`, samples are columns and features are rows.}

\item{scale}{Logical. Divide features by their standard deviation.}

\item{center}{Logical. Subtract features by their average.}

\item{metric}{A function that return a object of class "dist".}

\item{ndim}{Integer. Dimensions of the embedded space.}

\item{maxit}{The maximum number of iterations.}
}
\value{
A matrix of coordinates with samples as rows.
}
\description{
Non-metric Multidimensional Scaling (dimension reduction)
}
\examples{
data("bulkLogCounts")
NMDSproj<-NMDS(bulkLogCounts)
proj2d(NMDSproj)
}
