% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrix.R
\name{normDeseq}
\alias{normDeseq}
\title{Quick DESeq2 normalization}
\usage{
normDeseq(countMatrix)
}
\arguments{
\item{countMatrix}{A raw count table with rows as genes}
}
\value{
A normalized count table.
}
\description{
Quick DESeq2 normalization
}
\examples{
data("geneLengthGRCh38")
countMat<-t(sapply(vector("numeric",length = length(geneLengthGRCh38)),function(x){
    MASS::rnegbin(10,theta = abs(rnorm(1,mean = 10,sd = 20)),mu = abs(rnorm(1,mean = 10,sd = 20)))
}));rownames(countMat)<-names(geneLengthGRCh38)
normDeseq(countMat)
}
