% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oob-package.R
\docType{package}
\name{oob-package}
\alias{oob}
\alias{oob-package}
\title{oob: oob: a toolbox for analyzing OMIC data, Out Of Bounds from any black box}
\description{
A toolbox for analyzing OMIC data, Out Of Bounds from any black box by avoiding the use of complex objects. This package provide function for analyzing OMIC data, with a focus on plots and bulk/single-cell RNA-Seq. In addition of providing diverse wrappers and utilities to ease the use of R with large datsets, it contains several unique features. Among them, the TRIMAP dimension reduction implementated with reticulate and tools for enriching intersection of gene lists.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DimitriMeistermann/oob}
  \item Report bugs at \url{https://github.com/DimitriMeistermann/oob/issues}
}

}
\author{
\strong{Maintainer}: Dimitri Meistermann \email{dimitri.meistermann@helsinki.fi} (\href{https://orcid.org/0000-0001-6614-2325}{ORCID})

}
\keyword{internal}
