% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oob-package.R
\docType{package}
\name{oob-package}
\alias{oob}
\alias{oob-package}
\title{oob: oob: a toolbox for analyzing OMIC data, Out Of Bounds from any black box.}
\description{
A toolbox for analyzing OMIC data, Out Of Bounds from any black box.
}
\keyword{internal}
