% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrix.R
\name{oobFastMNN}
\alias{oobFastMNN}
\title{Execute a fastMNN and rescale the counts.}
\usage{
oobFastMNN(logCounts, batch, k, returnRescale = TRUE, ...)
}
\arguments{
\item{logCounts}{A matrix of numeric (in the RNA-Seq context, log counts).}

\item{batch}{A vector or factor specifying the batch of origin for all cells}

\item{k}{An integer scalar specifying the number of nearest neighbors to consider when identifying MNNs.}

\item{returnRescale}{Logical. Use reScale on the output so the dymaic range after bacth correction of genes is the same than before.}

\item{...}{Other parameters passed to fastMNN.}
}
\value{
A matrix of corrected count table.
}
\description{
Use fastMNN fron the package batchelor, but return a matrix. Rescale the counts so the range of each gene remains the same after the transformation.
}
\examples{
data("geneLengthGRCh38")
library(MASS)
countMat<-sapply(vector("numeric",length = 100),function(x){
 	c(rnegbin(10,mu = 50,theta = 5), rnegbin(10,mu = 10,theta = 5))
}) |> t(); countMat<-log2(countMat+1)
heatmap.DM(countMat)
correctedMat<-oobFastMNN(countMat,batch = c(rep(1,10),rep(2,10)), k=5) #warning because of the small matrix
heatmap.DM(correctedMat)
}
