% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{proj2dWideColor}
\alias{proj2dWideColor}
\title{Wrapper for proj2d with highly contrasted color scale.}
\usage{
proj2dWideColor(
  coord,
  axis = 1:2,
  colorBy = NULL,
  pointSize = 3,
  plotText = FALSE,
  main = NULL,
  alpha = 9/10,
  ellipse = FALSE,
  emph = NULL,
  returnGraph = FALSE,
  legendTitle = "Values",
  axis.names = NULL,
  na.color = "grey50",
  na.bg = TRUE,
  funAutoColorScale = ggplotColours,
  ...
)
}
\description{
Wrapper for proj2d with highly contrasted color scale.
}
\examples{
data(iris)
proj2dWideColor(iris,colorBy = iris$Petal.Length)
}
