% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{richUpset}
\alias{richUpset}
\title{Upset plot with additional enrichment values.}
\usage{
richUpset(featurePerGroupList, universe = NULL)
}
\arguments{
\item{featurePerGroupList}{A list of sets (vector of charachter)}

\item{universe}{NULL or vector of Character. The entire list of features (Universal set).}
}
\value{
Plot in the current graphical device.
}
\description{
Upset plot with additional enrichment values.
}
\examples{
lt = list(set1 = sample(letters, 5),
					set2 = sample(letters, 10),
					set3 = sample(letters, 15)
					)
lt$set4 <- unique(c(lt$set1,lt$set2))

richUpset(lt, universe = letters)
}
