% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setRandState}
\alias{setRandState}
\title{Set the precise random seed state}
\usage{
setRandState(state)
}
\arguments{
\item{state}{Object saved by getRandState}
}
\value{
NULL, set the random seed state
}
\description{
Set the precise random seed state
}
\examples{
randSate<-getRandState()
rnorm(5)
setRandState(randSate)
rnorm(5)
}
\seealso{
getRandState
}
