% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMatrix.R
\name{TRIMAP}
\alias{TRIMAP}
\title{TriMap dimension reduction.}
\usage{
TRIMAP(
  data,
  n_dims = 2,
  transpose = TRUE,
  n_inliers = 10,
  n_outliers = 5,
  apply_pca = TRUE,
  n_iters = 400,
  knn_tuple = NULL
)
}
\arguments{
\item{data}{A matrix of numeric (in the RNA-Seq context, log counts).}

\item{n_dims}{Integer. Dimensions of the embedded space.}

\item{transpose}{Logical. If \code{transpose}, samples are columns and features are rows.}

\item{n_inliers}{Number of nearest neighbors for forming the nearest neighbor triplets.}

\item{n_outliers}{Number of outliers for forming the nearest neighbor triplets.}

\item{apply_pca}{Reduce the number of dimensions of the data to 100 if necessary before applying the nearest-neighbor search.}

\item{n_iters}{Number of iterations.}

\item{knn_tuple}{Use the precomputed nearest-neighbors information in form of a tuple (knn_nbrs, knn_distances).}
}
\value{
A matrix of coordinates with samples as rows.
}
\description{
TriMap dimension reduction.
}
\examples{
data(iris)
irisTrimap<-TRIMAP(iris[,1:4],transpose=FALSE)
proj2d(irisTrimap,colorBy = iris$Species)
}
